refactor(rendering): decompose PipelineOwner god-file into cohesive submodules
Split pipeline/owner/mod.rs (4218 LOC, ~144 items) into 8 focused submodules:
- construction.rs  — PipelineOwner<Idle> constructors + run_frame orchestrator
- accessors.rs     — phase-generic accessors, insertion, dirty tracking, hit-test
- layout.rs        — PipelineOwner<Layout>: run_layout + deferred mutation apply
- compositing.rs   — PipelineOwner<Compositing>: compositing-bits walk
- paint.rs         — PipelineOwner<PaintPhase>: paint subtree walk + FragmentComposer
- semantics.rs     — PipelineOwner<Semantics>: run_semantics
- query.rs         — memoizing intrinsic/dry-layout walks + QuerySlot
- diagnostics.rs   — Diagnosticable-backed tree introspection

subtree_arena.rs (all pipeline unsafe, 1704 LOC) is untouched except cosmetic
rustfmt re-flow. rebind_phase, the struct definition, Debug/Default impls,
and all tests remain in mod.rs. Public API is byte-identical.

Gates: check clean, 588/588 nextest, 446/446 flui-view, fmt clean,
clippy -D warnings clean, port-check all 20 triggers clean.
