+++ crates/flui-rendering/src/pipeline/owner/mod.rs
@@ -7,41 +7,37 @@
 //! doctest at the end of `pipeline/phase.rs`.
 
 mod subtree_arena;
-use subtree_arena::{SubtreeArena, child_flex_from_seeds, ensure_stack};
 
-use super::scheduler::{DirtyTracker, PhaseKind};
+mod accessors;
+mod compositing;
+mod construction;
+mod diagnostics;
+mod layout;
+mod paint;
+mod query;
+mod semantics;
 
 use std::{
     marker::PhantomData,
-    sync::atomic::{AtomicBool, AtomicU64, Ordering},
+    sync::atomic::{AtomicBool, AtomicU64},
 };
 
-use flui_foundation::{DiagnosticsNode, LayerId, RenderId};
-use flui_layer::{
-    ClipPathLayer, ClipRRectLayer, ClipRectLayer, Layer, LayerTree, OffsetLayer, OpacityLayer,
-    PictureLayer, TransformLayer,
-};
-use flui_painting::DisplayList;
-use flui_types::{Offset, Size};
+use flui_foundation::RenderId;
+use flui_layer::LayerTree;
 #[cfg(any(test, feature = "testing"))]
 use rustc_hash::FxHashMap;
-use rustc_hash::FxHashSet;
 
 #[cfg(any(test, feature = "testing"))]
 use crate::testing::parent_data::ParentDataSeed;
 
-use crate::{
-    constraints::{BoxConstraints, SliverConstraints, SliverGeometry},
-    context::{FragmentClip, FragmentOp, FragmentRecorder},
-    protocol::{BoxProtocol, MainAxisPosition, SliverProtocol},
-    storage::{RenderNode, RenderTree},
-};
+use crate::{constraints::BoxConstraints, storage::RenderTree};
 
 use super::{
-    dirty::DirtyNode,
-    handle::{DirtyKind, DirtyRequest, PipelineOwnerHandle},
+    deferred::DeferredMutations,
+    handle::{DirtyRequest, PipelineOwnerHandle},
     notifier::VisualUpdateNotifier,
-    phase::{Compositing, Idle, Layout, PaintPhase, PipelinePhase, Semantics},
+    phase::{Idle, PipelinePhase},
+    scheduler::DirtyTracker,
 };
 
 /// Default bounded capacity of the dirty-request channel between
@@ -163,7 +159,7 @@ pub struct PipelineOwner<Phase: PipelinePhase = Idle> {
     ///
     /// This is the Rust-native alternative to Flutter's
     /// `invokeLayoutCallback` which uses unsafe re-entrant mutation.
-    deferred_mutations: super::deferred::DeferredMutations,
+    deferred_mutations: DeferredMutations,
 
     /// Prototype handle held by the owner so `handle()` can clone it for
     /// each caller without re-allocating the channel. See
@@ -214,2749 +210,6 @@ impl Default for PipelineOwner<Idle> {
     }
 }
 
-// ============================================================================
-// Idle-only impl: constructors + orchestration
-// ============================================================================
-
-impl PipelineOwner<Idle> {
-    /// Creates a new pipeline owner in the [`Idle`] phase with the
-    /// default dirty-channel capacity (`DEFAULT_DIRTY_CHANNEL_CAPACITY`,
-    /// 256).
-    pub fn new() -> Self {
-        Self::new_with_capacity(DEFAULT_DIRTY_CHANNEL_CAPACITY)
-    }
-
-    /// Creates a new pipeline owner in the [`Idle`] phase with a custom
-    /// dirty-channel capacity. Use this when the default 256 doesn't match
-    /// the producer profile.
-    pub fn new_with_capacity(dirty_channel_capacity: usize) -> Self {
-        let notifier = std::sync::Arc::new(parking_lot::RwLock::new(VisualUpdateNotifier::new()));
-        let (handle, dirty_rx) =
-            PipelineOwnerHandle::new_pair(dirty_channel_capacity, std::sync::Arc::clone(&notifier));
-        let scheduler = DirtyTracker::new(std::sync::Arc::clone(&notifier));
-        Self {
-            id: PIPELINE_ID_COUNTER.fetch_add(1, Ordering::Relaxed),
-            render_tree: RenderTree::new(),
-            root_id: None,
-            notifier,
-            scheduler,
-            root_constraints: None,
-            semantics_enabled: AtomicBool::new(false),
-            last_layer_tree: None,
-            device_pixel_ratio: 1.0,
-            deferred_mutations: super::deferred::DeferredMutations::new(),
-            handle,
-            dirty_rx,
-            #[cfg(any(test, feature = "testing"))]
-            parent_data_seeds: FxHashMap::default(),
-            _phase: PhantomData,
-        }
-    }
-
-    /// Records harness parent metadata for `child_id`, cloned into the
-    /// transient child slots before each layout walk.
-    ///
-    /// A second call for the same `child_id` replaces the previous seed.
-    #[cfg(any(test, feature = "testing"))]
-    pub fn seed_parent_data(&mut self, child_id: RenderId, seed: ParentDataSeed) {
-        self.parent_data_seeds.insert(child_id, seed);
-    }
-
-    /// Creates a new pipeline owner with callbacks in the [`Idle`] phase.
-    pub fn with_callbacks<F, G, H>(
-        on_need_visual_update: Option<F>,
-        on_semantics_owner_created: Option<G>,
-        on_semantics_owner_disposed: Option<H>,
-    ) -> Self
-    where
-        F: Fn() + Send + Sync + 'static,
-        G: Fn() + Send + Sync + 'static,
-        H: Fn() + Send + Sync + 'static,
-    {
-        let mut notifier = VisualUpdateNotifier::new();
-        if let Some(f) = on_need_visual_update {
-            notifier.set_need_visual_update(f);
-        }
-        if let Some(f) = on_semantics_owner_created {
-            notifier.set_semantics_owner_created(f);
-        }
-        if let Some(f) = on_semantics_owner_disposed {
-            notifier.set_semantics_owner_disposed(f);
-        }
-        let notifier = std::sync::Arc::new(parking_lot::RwLock::new(notifier));
-        let (handle, dirty_rx) = PipelineOwnerHandle::new_pair(
-            DEFAULT_DIRTY_CHANNEL_CAPACITY,
-            std::sync::Arc::clone(&notifier),
-        );
-        let scheduler = DirtyTracker::new(std::sync::Arc::clone(&notifier));
-        Self {
-            id: PIPELINE_ID_COUNTER.fetch_add(1, Ordering::Relaxed),
-            render_tree: RenderTree::new(),
-            root_id: None,
-            notifier,
-            scheduler,
-            root_constraints: None,
-            semantics_enabled: AtomicBool::new(false),
-            last_layer_tree: None,
-            device_pixel_ratio: 1.0,
-            deferred_mutations: super::deferred::DeferredMutations::new(),
-            handle,
-            dirty_rx,
-            #[cfg(any(test, feature = "testing"))]
-            parent_data_seeds: FxHashMap::default(),
-            _phase: PhantomData,
-        }
-    }
-
-    /// Transitions an idle pipeline into the [`Layout`] phase.
-    ///
-    /// Consumes `self`; once transitioned out of `Idle`, the legacy
-    /// idle-only API (constructors, `run_frame`) is no longer reachable
-    /// until you return through [`finish`](PipelineOwner::<Semantics>::finish).
-    #[must_use]
-    pub fn into_layout(self) -> PipelineOwner<Layout> {
-        rebind_phase(self)
-    }
-
-    // ========================================================================
-    // Full-frame orchestrator (Mythos Step 7)
-    // ========================================================================
-
-    /// Runs a full frame: layout -> compositing-bits -> paint -> semantics.
-    /// Consumes `self`, returns the owner back at [`Idle`] plus a
-    /// [`RenderResult`](crate::RenderResult) indicating whether the frame produced a layer
-    /// tree or failed mid-phase.
-    ///
-    /// The phase transitions are the load-bearing mechanism here -- each
-    /// `run_*` method lives only on its matching phase's impl block, so
-    /// the type system enforces the ordering. There is no runtime branch
-    /// that could call `run_paint` before `run_layout`.
-    ///
-    /// # Mythos Step 12 -- error handling
-    ///
-    /// If any phase returns [`crate::error::RenderError`] (most notably
-    /// [`crate::error::RenderError::Poisoned`] from a panicking render
-    /// object), the in-flight frame is dropped, the owner is returned at
-    /// [`Idle`] (no in-flight layer tree), and the second element of the
-    /// tuple is `Err(...)`. The owner is **always** usable for a
-    /// subsequent frame on the success and error paths alike.
-    #[must_use = "dropping the returned PipelineOwner<Idle> discards the pipeline handle; thread it back into the next frame"]
-    pub fn run_frame(
-        mut self,
-    ) -> (
-        PipelineOwner<Idle>,
-        crate::error::RenderResult<Option<LayerTree>>,
-    ) {
-        // Observe cross-thread dirty requests (RepaintHandle /
-        // PipelineOwnerHandle producers) before any phase runs — an
-        // async decode that finished while the app idled lands in this
-        // frame, not never.
-        self.drain_pending_dirty();
-
-        // Layout
-        let mut owner = self.into_layout();
-        if let Err(e) = owner.run_layout() {
-            return (owner.into_idle(), Err(e));
-        }
-
-        // Compositing
-        let mut owner = owner.into_compositing();
-        if let Err(e) = owner.run_compositing() {
-            return (owner.into_idle(), Err(e));
-        }
-
-        // Paint
-        let mut owner = owner.into_paint();
-        if let Err(e) = owner.run_paint() {
-            return (owner.into_idle(), Err(e));
-        }
-
-        // Semantics
-        let mut owner = owner.into_semantics();
-        if let Err(e) = owner.run_semantics() {
-            // Semantics phase has no `into_idle` because the transition
-            // to <Idle> goes via `finish`. Use `finish` to recover the
-            // owner for the error path -- the layer tree from the paint
-            // phase is discarded on error to keep the invariant "Err =>
-            // no layer tree".
-            return (owner.finish(), Err(e));
-        }
-
-        let layer_tree = owner.take_layer_tree();
-        (owner.finish(), Ok(layer_tree))
-    }
-}
-
-// ============================================================================
-// Phase-agnostic accessors / setters / insertion (Mythos Step 7)
-// ============================================================================
-//
-// These methods are pure data access or side-effect-free notifier wiring.
-// They are valid in any phase: the borrow checker still gates `&mut self`
-// against the type-state transitions, but the methods themselves don't
-// care which phase the owner is in.
-
-impl<Phase: PipelinePhase> PipelineOwner<Phase> {
-    /// Returns the unique identifier for this pipeline owner.
-    #[inline]
-    pub fn id(&self) -> u64 {
-        self.id
-    }
-
-    /// Sets the callback for when a visual update is needed.
-    pub fn set_on_need_visual_update<F>(&mut self, callback: F)
-    where
-        F: Fn() + Send + Sync + 'static,
-    {
-        self.notifier.write().set_need_visual_update(callback);
-    }
-
-    /// Sets the callback for when semantics owner is created.
-    pub fn set_on_semantics_owner_created<F>(&mut self, callback: F)
-    where
-        F: Fn() + Send + Sync + 'static,
-    {
-        self.notifier.write().set_semantics_owner_created(callback);
-    }
-
-    /// Sets the callback for when semantics owner is disposed.
-    pub fn set_on_semantics_owner_disposed<F>(&mut self, callback: F)
-    where
-        F: Fn() + Send + Sync + 'static,
-    {
-        self.notifier.write().set_semantics_owner_disposed(callback);
-    }
-
-    /// Requests a visual update.
-    ///
-    /// Called by render objects when they need to be re-rendered.
-    pub fn request_visual_update(&self) {
-        self.notifier.read().fire_need_visual_update();
-    }
-
-    // ========================================================================
-    // Cross-thread mark-dirty handle (Mythos Step 8)
-    // ========================================================================
-
-    /// Returns a clone of the cross-thread mark-dirty handle.
-    ///
-    /// Each clone is its own `Sender` over the same bounded channel; sends
-    /// from different threads do not block each other. Backpressure
-    /// surfaces as [`super::handle::SendError::ChannelFull`].
-    #[inline]
-    pub fn handle(&self) -> PipelineOwnerHandle {
-        self.handle.clone()
-    }
-
-    /// Binds a [`RepaintHandle`](crate::pipeline::RepaintHandle) to a
-    /// live render object — the
-    /// capability async producers (image decodes, arriving assets) use
-    /// to repaint that node from any thread, with the platform woken on
-    /// every request.
-    ///
-    /// `None` for a stale/foreign id. Once the node is later removed,
-    /// the returned handle degrades to a silent no-op (generational id:
-    /// the drain drops requests whose generation died).
-    pub fn repaint_handle(&self, id: RenderId) -> Option<crate::pipeline::RepaintHandle> {
-        let depth = self.render_tree.get(id)?.depth() as usize;
-        Some(crate::pipeline::RepaintHandle::new(
-            self.handle.clone(),
-            id,
-            depth,
-        ))
-    }
-
-    /// Drains the pending dirty-request channel into the scheduler's
-    /// dirty sets.
-    ///
-    /// Called at phase boundaries by the typestate transitions; producers
-    /// (background asset loaders, async work) write into the channel via
-    /// [`PipelineOwnerHandle::request_mark_dirty`] and the owner observes
-    /// them on the next frame. Non-blocking; processes every request
-    /// available at the time of call and returns the count drained.
-    pub fn drain_pending_dirty(&mut self) -> usize {
-        let mut drained = 0;
-        while let Ok(req) = self.dirty_rx.try_recv() {
-            drained += 1;
-            // Generation-validated replay through the SAME mark paths
-            // local callers use: a stale id (the node died after the
-            // producer captured its handle) falls out silently, the
-            // layout mark walks to its relayout boundary, compositing
-            // keeps the queue⇒flag invariant, and every path carries
-            // its own dedup. The request's depth is advisory — the live
-            // node's depth is authoritative.
-            //
-            // Pre-fix this pushed raw queue entries: a layout request
-            // for a non-boundary node became a bogus dirty ROOT, a
-            // compositing request skipped the flag (the silent-loss
-            // footgun `add_node_needing_compositing_bits_update`
-            // exists to prevent), and dead ids were replayed verbatim.
-            let Some(node) = self.render_tree.get(req.id) else {
-                tracing::trace!(?req, "drain_pending_dirty: stale id, dropped");
-                continue;
-            };
-            let depth = node.depth() as usize;
-            match req.kind {
-                DirtyKind::Layout => self.mark_needs_layout(req.id),
-                DirtyKind::Compositing => {
-                    self.add_node_needing_compositing_bits_update(req.id, depth);
-                }
-                DirtyKind::Paint => self.add_node_needing_paint(req.id, depth),
-                DirtyKind::Semantics => self.add_node_needing_semantics(req.id, depth),
-            }
-        }
-        drained
-    }
-
-    /// Returns the root render object ID.
-    pub fn root_id(&self) -> Option<RenderId> {
-        self.root_id
-    }
-
-    /// Sets the root render object ID.
-    pub fn set_root_id(&mut self, id: Option<RenderId>) {
-        self.root_id = id;
-    }
-
-    /// Returns the constraints to apply to the root render object on
-    /// the next layout pass (if no cached constraints yet).
-    ///
-    /// **D-block PR-A1 U23:** see [`Self::set_root_constraints`].
-    #[inline]
-    pub fn root_constraints(&self) -> Option<BoxConstraints> {
-        self.root_constraints
-    }
-
-    /// Sets the constraints to apply to the root render object on the
-    /// next layout pass when no cached constraints exist yet
-    /// (first-frame initialization).
-    ///
-    /// **D-block PR-A1 U23:** the binding layer (`flui-view` /
-    /// `flui-app` / `flui-hot-reload`) calls this once after
-    /// constructing the pipeline + before the first `run_frame`
-    /// invocation. On subsequent frames the root's cached
-    /// `state.constraints()` (post-layout) supersedes this field; the
-    /// fallback only fires on the very first layout pass.
-    ///
-    /// Pass `None` to clear (e.g., when the binding wants to defer to
-    /// a yet-unmounted root render object that supplies its own
-    /// constraints via `RootRenderElement::mount`).
-    ///
-    /// # Auto-schedules root relayout (PR #148 review fix)
-    ///
-    /// When `Some(_)` is passed AND `root_id` is set AND the new
-    /// constraints differ from the prior value, this method also
-    /// calls [`Self::mark_needs_layout`] on the root id so the
-    /// next `run_layout` invocation picks up the change. This
-    /// avoids the silent-no-relayout footgun the prior shape had —
-    /// the binding no longer needs to call `mark_needs_layout`
-    /// separately after `set_root_constraints`.
-    ///
-    /// Setting to the SAME constraints value (or to `None`) does
-    /// NOT mark dirty — those cases either don't change the layout
-    /// result or are explicit clears that the caller manages
-    /// independently.
-    pub fn set_root_constraints(&mut self, constraints: Option<BoxConstraints>) {
-        let changed = constraints.is_some() && constraints != self.root_constraints;
-        self.root_constraints = constraints;
-        if changed && let Some(root_id) = self.root_id {
-            self.mark_needs_layout(root_id);
-        }
-    }
-
-    /// Returns a reference to the render tree.
-    pub fn render_tree(&self) -> &RenderTree {
-        &self.render_tree
-    }
-
-    /// Returns a mutable reference to the render tree.
-    pub fn render_tree_mut(&mut self) -> &mut RenderTree {
-        &mut self.render_tree
-    }
-
-    /// Returns a reference to the layer tree from the last paint phase.
-    pub fn layer_tree(&self) -> Option<&LayerTree> {
-        self.last_layer_tree.as_ref()
-    }
-
-    /// Takes the layer tree from the last paint phase.
-    ///
-    /// This removes the layer tree from the pipeline owner, returning ownership
-    /// to the caller. Useful for passing to the compositor.
-    ///
-    /// Phase-agnostic: works in any phase. `run_frame` calls this on
-    /// `<Semantics>` to extract the layer tree before transitioning back to
-    /// `<Idle>`.
-    pub fn take_layer_tree(&mut self) -> Option<LayerTree> {
-        self.last_layer_tree.take()
-    }
-
-    /// Device pixel ratio threaded into every paint pass.
-    pub fn device_pixel_ratio(&self) -> f32 {
-        self.device_pixel_ratio
-    }
-
-    /// Removes the subtree rooted at `id` — THE dispose site.
-    ///
-    /// Removal is where owner-side state dies (the inversion of the
-    /// "Drop will handle it" idea): a `Drop` impl has no
-    /// `&PipelineOwner`, so it cannot evict dirty-queue entries — only
-    /// this method can. Order:
-    ///
-    /// 1. collect the subtree's ids;
-    /// 2. evict every id from ALL dirty queues (live + mid-phase) so
-    ///    no phase walks a freed slot's stale entry;
-    /// 3. cascade-remove the nodes (each freed slot's generation bumps
-    ///    — outstanding ids go stale, D2);
-    /// 4. clear `root_id` when the root itself was removed.
-    ///
-    /// `Drop` on render objects remains strictly node-local (decoded
-    /// images, shaped text, GPU handles).
-    ///
-    /// Returns the number of nodes removed. Must not run mid-phase —
-    /// the layout/paint walks hold raw borrows into the slab.
-    pub fn remove_render_object(&mut self, id: RenderId) -> usize {
-        debug_assert!(
-            !self.scheduler.debug_doing_layout() && !self.scheduler.debug_doing_paint(),
-            "remove_render_object during an active layout/paint phase — \
-             the walks hold borrows into the slab; defer removal to \
-             between-frame work",
-        );
-
-        let subtree = self.render_tree.collect_subtree_ids(id);
-        if subtree.is_empty() {
-            return 0;
-        }
-        let removed: FxHashSet<RenderId> = subtree.iter().copied().collect();
-        self.scheduler.evict(&removed);
-
-        let count = self.render_tree.remove_recursive(id);
-        if self.root_id == Some(id) {
-            self.root_id = None;
-        }
-        tracing::debug!(?id, count, "remove_render_object: subtree disposed");
-        count
-    }
-
-    /// Hit-tests the render tree at `position` (root-local
-    /// coordinates), appending leaf-first entries to `result`.
-    ///
-    /// The walk mirrors the paint walk's shape: per node, the
-    /// protocol blanket wraps a driver-supplied child-recursion
-    /// callback in the typed, arity-gated hit-test context and calls
-    /// the object's `hit_test`. `None` child overrides resolve to the
-    /// child's laid-out `RenderState.offset` — parents no longer
-    /// mirror offsets in their own fields to hit-test children.
-    ///
-    /// Returns whether anything was hit. O(visited nodes) average;
-    /// worst case O(tree) when nothing claims the position.
-    pub fn hit_test(&self, position: Offset, result: &mut crate::hit_testing::HitTestResult) -> bool
-    where
-        // The child-recursion callback must be Send + Sync (ctx trait
-        // bounds, U19 inheritance); it captures &self, so the phase
-        // marker must be Sync. Every phase is a ZST — always satisfied
-        // at call sites, spelled out for the generic impl.
-        Phase: Sync,
-    {
-        let Some(root_id) = self.root_id else {
-            return false;
-        };
-        self.hit_test_subtree(root_id, position, result)
-    }
-
-    fn hit_test_subtree(
-        &self,
-        id: RenderId,
-        position: Offset,
-        result: &mut crate::hit_testing::HitTestResult,
-    ) -> bool
-    where
-        Phase: Sync,
-    {
-        ensure_stack(|| self.hit_test_subtree_impl(id, position, result))
-    }
-
-    /// Body of [`Self::hit_test_subtree`]; split out so every
-    /// recursion level enters through the [`ensure_stack`] probe.
-    fn hit_test_subtree_impl(
-        &self,
-        id: RenderId,
-        position: Offset,
-        result: &mut crate::hit_testing::HitTestResult,
-    ) -> bool
-    where
-        Phase: Sync,
-    {
-        let Some(node) = self.render_tree.get(id) else {
-            return false;
-        };
-        let Some(entry) = node.as_box() else {
-            if node.as_sliver().is_some() {
-                let sliver_position = Self::sliver_hit_position_from_offset(node, position);
-                return self.hit_test_sliver_subtree(id, sliver_position, result);
-            }
-            return false;
-        };
-        let children: Vec<RenderId> = node.children().to_vec();
-        let render_object = entry.render_object();
-        // Box bounds gate size, resolved from RenderState (geometry's
-        // sole owner — 2B field dedup); threaded into hit_test_raw so the
-        // default gate reads `ctx.is_within_own_size()`.
-        let own_size = entry.state().geometry().unwrap_or(flui_types::Size::ZERO);
-
-        // Push the node's hit-test transform onto the HitTestResult
-        // stack BEFORE recursing, so child entries captured during
-        // hit_test_raw see the correct accumulated transform. The
-        // transform stays on the stack until after the parent entry
-        // is added (Flutter parity: pushTransform in hitTest). The hook
-        // gets `own_size` from RenderState for alignment-relative origins.
-        let hit_transform = render_object.hit_test_transform(own_size);
-        let has_transform = hit_transform.is_some();
-        if let Some(t) = hit_transform {
-            result.push_transform(t);
-        }
-
-        let mut hit_child = |index: usize, override_pos: Option<Offset>| -> bool {
-            let Some(&child_id) = children.get(index) else {
-                return false;
-            };
-            let Some(child_node) = self.render_tree.get(child_id) else {
-                return false;
-            };
-            if child_node.as_sliver().is_some() {
-                // Explicit positions are already child-local; the layout-offset
-                // fallback starts from the child's physical paint offset.
-                return match override_pos {
-                    Some(position) => {
-                        let child_position =
-                            Self::sliver_hit_position_from_offset(child_node, position);
-                        self.hit_test_sliver_subtree(child_id, child_position, result)
-                    }
-                    None => {
-                        if !Self::sliver_child_is_visible(child_node) {
-                            return false;
-                        }
-                        let child_offset = child_node.offset();
-                        result.with_paint_offset(child_offset, |result| {
-                            let child_position = Self::sliver_hit_position_from_paint_offset(
-                                child_node,
-                                position - child_offset,
-                            );
-                            self.hit_test_sliver_subtree(child_id, child_position, result)
-                        })
-                    }
-                };
-            }
-            match override_pos {
-                Some(child_position) => self.hit_test_subtree(child_id, child_position, result),
-                None => {
-                    let child_offset = child_node.offset();
-                    result.with_paint_offset(child_offset, |result| {
-                        self.hit_test_subtree(child_id, position - child_offset, result)
-                    })
-                }
-            }
-        };
-
-        let hit = render_object.hit_test_raw(position, children.len(), own_size, &mut hit_child);
-        if hit {
-            // Leaf-first path: children pushed their entries during
-            // the callback above; the ancestor follows. The transform
-            // is still on the stack, so this entry captures it.
-            //
-            // A render object that listens for pointer events (RenderListener)
-            // advertises a handler here; it rides on the entry so
-            // `HitTestResult::dispatch` can invoke it. Additive: every other
-            // render object returns `None`, leaving today's handler-less entry.
-            let entry = crate::hit_testing::HitTestEntry::new(id);
-            let entry = match render_object.pointer_event_handler() {
-                Some(handler) => entry.handler(handler),
-                None => entry,
-            };
-            result.add(entry);
-        }
-
-        if has_transform {
-            result.pop_transform();
-        }
-
-        hit
-    }
-
-    fn sliver_hit_position_from_offset(node: &RenderNode, position: Offset) -> MainAxisPosition {
-        let axis_direction = node
-            .as_sliver()
-            .and_then(|entry| entry.state().constraints())
-            .map(|constraints| constraints.axis_direction);
-
-        match axis_direction {
-            Some(
-                flui_types::layout::AxisDirection::LeftToRight
-                | flui_types::layout::AxisDirection::RightToLeft,
-            ) => MainAxisPosition::from_horizontal_offset(position),
-            _ => MainAxisPosition::from_vertical_offset(position),
-        }
-    }
-
-    fn sliver_hit_position_from_paint_offset(
-        node: &RenderNode,
-        position: Offset,
-    ) -> MainAxisPosition {
-        let Some(entry) = node.as_sliver() else {
-            return MainAxisPosition::from_vertical_offset(position);
-        };
-        let Some(constraints) = entry.state().constraints() else {
-            return MainAxisPosition::from_vertical_offset(position);
-        };
-
-        let (raw_main, cross_axis) = match constraints.axis_direction {
-            flui_types::layout::AxisDirection::LeftToRight
-            | flui_types::layout::AxisDirection::RightToLeft => {
-                (position.dx.get(), position.dy.get())
-            }
-            flui_types::layout::AxisDirection::TopToBottom
-            | flui_types::layout::AxisDirection::BottomToTop => {
-                (position.dy.get(), position.dx.get())
-            }
-        };
-        let effective_axis_direction = match constraints.growth_direction {
-            crate::constraints::GrowthDirection::Forward => constraints.axis_direction,
-            crate::constraints::GrowthDirection::Reverse => constraints.axis_direction.opposite(),
-        };
-        let main_axis = if effective_axis_direction.is_reversed() {
-            entry
-                .state()
-                .geometry()
-                .map_or(raw_main, |geometry| geometry.paint_extent - raw_main)
-        } else {
-            raw_main
-        };
-
-        MainAxisPosition::new(main_axis, cross_axis)
-    }
-
-    fn sliver_child_is_visible(node: &RenderNode) -> bool {
-        node.as_sliver()
-            .and_then(|entry| entry.state().geometry())
-            .is_some_and(|geometry| geometry.visible)
-    }
-
-    fn sliver_hit_position_minus_paint_offset(
-        parent_constraints: &SliverConstraints,
-        parent_geometry: &SliverGeometry,
-        child_node: &RenderNode,
-        position: MainAxisPosition,
-        offset: Offset,
-    ) -> MainAxisPosition {
-        let Some(child_entry) = child_node.as_sliver() else {
-            return position;
-        };
-        let Some(child_constraints) = child_entry.state().constraints() else {
-            return position;
-        };
-        let Some(child_geometry) = child_entry.state().geometry() else {
-            return position;
-        };
-
-        let (offset_main, offset_cross) = match child_constraints.axis_direction {
-            flui_types::layout::AxisDirection::LeftToRight
-            | flui_types::layout::AxisDirection::RightToLeft => (offset.dx.get(), offset.dy.get()),
-            flui_types::layout::AxisDirection::TopToBottom
-            | flui_types::layout::AxisDirection::BottomToTop => (offset.dy.get(), offset.dx.get()),
-        };
-        let parent_physical_main = if parent_constraints
-            .growth_direction
-            .apply_to_axis_direction(parent_constraints.axis_direction)
-            .is_reversed()
-        {
-            parent_geometry.paint_extent - position.main_axis
-        } else {
-            position.main_axis
-        };
-        let child_physical_main = parent_physical_main - offset_main;
-        let child_main = if child_constraints
-            .growth_direction
-            .apply_to_axis_direction(child_constraints.axis_direction)
-            .is_reversed()
-        {
-            child_geometry.paint_extent - child_physical_main
-        } else {
-            child_physical_main
-        };
-
-        MainAxisPosition::new(child_main, position.cross_axis - offset_cross)
-    }
-
-    fn box_hit_offset_from_sliver_position(
-        constraints: &SliverConstraints,
-        geometry: &SliverGeometry,
-        child_size: Size,
-        position: MainAxisPosition,
-        offset: Offset,
-    ) -> Offset {
-        let right_way_up = crate::constraints::right_way_up(
-            constraints.axis_direction,
-            constraints.growth_direction,
-        );
-
-        let (paint_main, paint_cross, child_main_extent) = match constraints.axis_direction {
-            flui_types::layout::AxisDirection::LeftToRight
-            | flui_types::layout::AxisDirection::RightToLeft => {
-                (offset.dx.get(), offset.dy.get(), child_size.width.get())
-            }
-            flui_types::layout::AxisDirection::TopToBottom
-            | flui_types::layout::AxisDirection::BottomToTop => {
-                (offset.dy.get(), offset.dx.get(), child_size.height.get())
-            }
-        };
-        let child_main_axis_position = if right_way_up {
-            paint_main
-        } else {
-            geometry.paint_extent - child_main_extent - paint_main
-        };
-        let mut local_main = position.main_axis - child_main_axis_position;
-        if !right_way_up {
-            local_main = child_main_extent - local_main;
-        }
-        let local_cross = position.cross_axis - paint_cross;
-
-        match constraints.axis_direction {
-            flui_types::layout::AxisDirection::LeftToRight
-            | flui_types::layout::AxisDirection::RightToLeft => Offset::new(
-                flui_types::geometry::px(local_main),
-                flui_types::geometry::px(local_cross),
-            ),
-            flui_types::layout::AxisDirection::TopToBottom
-            | flui_types::layout::AxisDirection::BottomToTop => Offset::new(
-                flui_types::geometry::px(local_cross),
-                flui_types::geometry::px(local_main),
-            ),
-        }
-    }
-
-    fn hit_test_sliver_subtree(
-        &self,
-        id: RenderId,
-        position: MainAxisPosition,
-        result: &mut crate::hit_testing::HitTestResult,
-    ) -> bool
-    where
-        Phase: Sync,
-    {
-        ensure_stack(|| self.hit_test_sliver_subtree_impl(id, position, result))
-    }
-
-    fn hit_test_sliver_subtree_impl(
-        &self,
-        id: RenderId,
-        position: MainAxisPosition,
-        result: &mut crate::hit_testing::HitTestResult,
-    ) -> bool
-    where
-        Phase: Sync,
-    {
-        let Some(node) = self.render_tree.get(id) else {
-            return false;
-        };
-        let Some(entry) = node.as_sliver() else {
-            return false;
-        };
-        let Some(geometry) = entry.state().geometry() else {
-            return false;
-        };
-        let Some(constraints) = entry.state().constraints() else {
-            return false;
-        };
-        if position.main_axis < 0.0
-            || position.main_axis >= geometry.hit_test_extent
-            || position.cross_axis < 0.0
-            || position.cross_axis >= constraints.cross_axis_extent
-        {
-            return false;
-        }
-        let children: Vec<RenderId> = node.children().to_vec();
-        let render_object = entry.render_object();
-        // Absolute paint size from RenderState — threaded into
-        // hit_test_raw for signature uniformity; the sliver hit gate is
-        // driver-owned (checked above), so the sliver context ignores it.
-        let own_size = entry.state().absolute_paint_size();
-
-        let mut hit_child = |index: usize, override_pos: Option<MainAxisPosition>| -> bool {
-            let Some(&child_id) = children.get(index) else {
-                return false;
-            };
-            let Some(child_node) = self.render_tree.get(child_id) else {
-                return false;
-            };
-            if child_node.as_sliver().is_some() {
-                // Explicit positions are already in child sliver coordinates.
-                // The layout-offset fallback starts from physical paint data.
-                let child_position = match override_pos {
-                    Some(position) => position,
-                    None => Self::sliver_hit_position_minus_paint_offset(
-                        constraints,
-                        &geometry,
-                        child_node,
-                        position,
-                        child_node.offset(),
-                    ),
-                };
-                self.hit_test_sliver_subtree(child_id, child_position, result)
-            } else if let Some(child_entry) = child_node.as_box() {
-                let Some(child_size) = child_entry.state().geometry() else {
-                    return false;
-                };
-                let child_position = Self::box_hit_offset_from_sliver_position(
-                    constraints,
-                    &geometry,
-                    child_size,
-                    override_pos.unwrap_or(position),
-                    child_node.offset(),
-                );
-                self.hit_test_subtree(child_id, child_position, result)
-            } else {
-                false
-            }
-        };
-
-        let hit = render_object.hit_test_raw(position, children.len(), own_size, &mut hit_child);
-        if hit {
-            result.add(crate::hit_testing::HitTestEntry::new(id));
-        }
-        hit
-    }
-
-    /// Sets the device pixel ratio for subsequent paint passes.
-    ///
-    /// Called by the platform binding on surface creation and DPI
-    /// change. Non-finite or non-positive values are rejected (kept at
-    /// the previous ratio) — a zero or NaN DPR poisons every shaped
-    /// glyph and snapped hairline downstream.
-    pub fn set_device_pixel_ratio(&mut self, dpr: f32) {
-        if dpr.is_finite() && dpr > 0.0 {
-            self.device_pixel_ratio = dpr;
-        } else {
-            tracing::warn!(
-                dpr,
-                "set_device_pixel_ratio: rejecting non-finite / non-positive \
-                 ratio; keeping {}",
-                self.device_pixel_ratio,
-            );
-        }
-    }
-
-    // ========================================================================
-    // RenderObject Insertion (with dirty tracking)
-    // ========================================================================
-
-    /// Inserts a render object into the tree and marks it as needing layout.
-    ///
-    /// This method:
-    /// 1. Inserts the render object into the RenderTree
-    /// 2. Adds the node to the dirty layout list (since new nodes need layout)
-    /// 3. Adds the node to the dirty paint list (since new nodes need paint)
-    ///
-    /// Use this instead of `render_tree_mut().insert()` to ensure proper dirty
-    /// tracking.
-    ///
-    /// # Returns
-    ///
-    /// The `RenderId` of the inserted node.
-    pub fn insert<P>(&mut self, render_object: Box<dyn crate::traits::RenderObject<P>>) -> RenderId
-    where
-        P: crate::protocol::Protocol,
-        crate::storage::RenderNode: From<Box<dyn crate::traits::RenderObject<P>>>,
-    {
-        use flui_tree::traits::TreeWrite;
-
-        // Convert to RenderNode using From impl (zero-cost, compile-time dispatch)
-        let node: crate::storage::RenderNode = render_object.into();
-        let id = self.render_tree.insert(node);
-        let depth = self.render_tree.depth(id).unwrap_or(0) as usize;
-
-        // PR-A2 U33: bootstrap IS_REPAINT_BOUNDARY flag from the
-        // render_object's static answer before the dirty pushes (so
-        // the compositing walk has accurate boundary info on first
-        // run_compositing).
-        self.bootstrap_repaint_boundary_flag(id);
-
-        // New nodes need layout and paint
-        self.add_node_needing_layout(id, depth);
-        self.add_node_needing_paint(id, depth);
-
-        id
-    }
-
-    /// Inserts a render object as a child and marks it as needing layout.
-    ///
-    /// This method:
-    /// 1. Inserts the render object as a child in the RenderTree
-    /// 2. Adds the node to the dirty layout list
-    /// 3. Adds the node to the dirty paint list
-    /// 4. Marks the parent as needing layout (since child structure changed)
-    ///
-    /// Use this instead of `render_tree_mut().insert_child()` to ensure proper
-    /// dirty tracking.
-    ///
-    /// # Arguments
-    ///
-    /// * `parent_id` - The parent node ID
-    /// * `render_object` - The render object to insert as child
-    ///
-    /// # Returns
-    ///
-    /// The `RenderId` of the inserted child, or `None` if parent doesn't exist.
-    pub fn insert_child_render_object(
-        &mut self,
-        parent_id: RenderId,
-        render_object: Box<dyn crate::traits::RenderObject<crate::protocol::BoxProtocol>>,
-    ) -> Option<RenderId> {
-        // Get parent depth before insertion
-        let parent_depth = self.render_tree.depth(parent_id)?;
-
-        // Insert child (using Box protocol)
-        let child_id = self
-            .render_tree
-            .insert_box_child(parent_id, render_object)?;
-        let child_depth = parent_depth + 1;
-
-        // PR-A2 U33: bootstrap IS_REPAINT_BOUNDARY flag from the
-        // child render_object's static answer before any compositing
-        // walk runs.
-        self.bootstrap_repaint_boundary_flag(child_id);
-
-        // Mark child as needing layout and paint
-        self.add_node_needing_layout(child_id, child_depth as usize);
-        self.add_node_needing_paint(child_id, child_depth as usize);
-
-        // Mark parent as needing layout (child structure changed)
-        self.add_node_needing_layout(parent_id, parent_depth as usize);
-
-        Some(child_id)
-    }
-
-    /// Inserts a raw `RenderNode` directly into the tree.
-    ///
-    /// This bypasses the `RenderObject<P>` trait requirement and is used for
-    /// special nodes like `RenderView` that manage their own layout/paint
-    /// lifecycle outside the standard protocol dispatch.
-    ///
-    /// # Returns
-    ///
-    /// The `RenderId` of the inserted node.
-    pub fn insert_render_node(&mut self, node: crate::storage::RenderNode) -> RenderId {
-        use flui_tree::traits::TreeWrite;
-
-        let id = self.render_tree.insert(node);
-        let depth = self.render_tree.depth(id).unwrap_or(0) as usize;
-
-        // PR-A2 U33: bootstrap IS_REPAINT_BOUNDARY flag (matches the
-        // `insert` / `insert_child_render_object` paths so every code
-        // path that adds nodes leaves the compositing flag in sync
-        // with the trait answer).
-        self.bootstrap_repaint_boundary_flag(id);
-
-        self.add_node_needing_layout(id, depth);
-        self.add_node_needing_paint(id, depth);
-
-        id
-    }
-
-    /// Sets the root render object and marks it as needing layout.
-    ///
-    /// This is a convenience method that:
-    /// 1. Inserts the render object
-    /// 2. Sets it as the root
-    /// 3. Ensures it's in the dirty lists
-    ///
-    /// # Returns
-    ///
-    /// The `RenderId` of the root node.
-    pub fn set_root_render_object(
-        &mut self,
-        render_object: Box<dyn crate::traits::RenderObject<crate::protocol::BoxProtocol>>,
-    ) -> RenderId {
-        let id = self.insert(render_object);
-        self.root_id = Some(id);
-        id
-    }
-
-    /// Bootstraps the `IS_REPAINT_BOUNDARY` storage flag from the render
-    /// object's static trait answer (`RenderObject::is_repaint_boundary()`).
-    ///
-    /// **D-block PR-A2 U33 (memo R26b).** Every node-insert path
-    /// ([`Self::insert`], [`Self::insert_child_render_object`],
-    /// [`Self::insert_render_node`]; [`Self::set_root_render_object`]
-    /// inherits via `insert`) calls this immediately after the tree
-    /// `insert` so the compositing-bits walk (U34) has accurate
-    /// boundary info via the storage flag from frame 1.
-    ///
-    /// **Current consumer scope:** the compositing-bits walk consults
-    /// `RenderNode::is_repaint_boundary_flag()`. The paint walk
-    /// (`paint_subtree`) still reads `render_object.is_repaint_boundary()`
-    /// directly — this matches Flutter parity (Flutter's `paint`
-    /// reads the `isRepaintBoundary` final getter, equivalent to our
-    /// trait answer; the bootstrap flag is the optimization target
-    /// for a later sweep that swaps the paint check too).
-    ///
-    /// Pre-U33 the storage flag was effectively `false` for every node
-    /// from the moment it entered the tree, which forced the
-    /// compositing walk to fall through to the trait answer
-    /// (`render_object().is_repaint_boundary()`) at every check site —
-    /// a virtual dispatch and a divergence risk if a future caller
-    /// flipped the flag dynamically.
-    ///
-    /// No-op if `id` is not present (defensive — every call site holds
-    /// a freshly-inserted id, but a stale id passes through silently
-    /// rather than panicking).
-    #[inline]
-    fn bootstrap_repaint_boundary_flag(&self, id: RenderId) {
-        if let Some(node) = self.render_tree.get(id) {
-            let is_boundary = node.is_repaint_boundary();
-            node.set_repaint_boundary_flag(is_boundary);
-        }
-    }
-
-    // ========================================================================
-    // Dirty Node Access (Flutter API)
-    // ========================================================================
-
-    // ========================================================================
-    // Deferred Mutations (re-entrant layout)
-    // ========================================================================
-
-    /// Enqueues a deferred mutation to be applied after the layout pass.
-    ///
-    /// During layout, render objects may need to add, remove, or update
-    /// children. But the layout walk holds `&mut` on the subtree, making
-    /// direct mutation impossible. This method collects mutations in a
-    /// queue that is drained after the pass completes.
-    ///
-    /// `logical_index`: if `Some(li)`, the pipeline stamps `li` into the
-    /// inserted child's parent-data after insertion (if the parent-data
-    /// type implements `crate::parent_data::LogicalIndexParentData`).
-    /// Pass `None` for non-lazy inserts.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// // During layout:
-    /// owner.defer_insert_box(parent_id, Box::new(new_child), None, None);
-    /// owner.defer_remove(parent_id, child_id);
-    /// owner.defer_update(target_id, Box::new(|obj| { /* mutate */ }));
-    /// ```
-    pub fn defer_insert_box(
-        &mut self,
-        parent_id: RenderId,
-        render_object: Box<dyn crate::protocol::RenderObject<BoxProtocol>>,
-        index: Option<usize>,
-        logical_index: Option<usize>,
-        initial_parent_data: Option<Box<dyn crate::parent_data::ParentData>>,
-    ) {
-        self.deferred_mutations.insert_box(
-            parent_id,
-            render_object,
-            index,
-            logical_index,
-            initial_parent_data,
-        );
-    }
-
-    /// Enqueues a deferred Sliver child insertion.
-    ///
-    /// `logical_index`: if `Some(li)`, stamps `li` into the child's
-    /// parent-data after insertion.  Pass `None` for non-lazy inserts.
-    ///
-    /// `initial_parent_data`: pre-built parent-data to install on the fresh
-    /// node immediately after insertion.  See `DeferredMutation::Insert`.
-    pub fn defer_insert_sliver(
-        &mut self,
-        parent_id: RenderId,
-        render_object: Box<dyn crate::protocol::RenderObject<SliverProtocol>>,
-        index: Option<usize>,
-        logical_index: Option<usize>,
-        initial_parent_data: Option<Box<dyn crate::parent_data::ParentData>>,
-    ) {
-        self.deferred_mutations.insert_sliver(
-            parent_id,
-            render_object,
-            index,
-            logical_index,
-            initial_parent_data,
-        );
-    }
-
-    /// Enqueues a deferred removal.
-    pub fn defer_remove(&mut self, parent_id: RenderId, child_id: RenderId) {
-        self.deferred_mutations.remove(parent_id, child_id);
-    }
-
-    /// Enqueues a deferred update (e.g., animation driving properties).
-    ///
-    /// The updater receives `&mut dyn Any` — the caller downcasts to
-    /// the concrete render object type.
-    pub fn defer_update(
-        &mut self,
-        target_id: RenderId,
-        updater: Box<dyn FnOnce(&mut dyn std::any::Any) + Send + Sync>,
-    ) {
-        self.deferred_mutations.update(target_id, updater);
-    }
-
-    /// Returns the number of pending deferred mutations.
-    pub fn deferred_mutation_count(&self) -> usize {
-        self.deferred_mutations.len()
-    }
-
-    /// Returns the nodes needing layout.
-    ///
-    /// These are relayout boundaries that need to be laid out in the next
-    /// layout phase.
-    #[inline]
-    pub fn nodes_needing_layout(&self) -> &[DirtyNode] {
-        self.scheduler.nodes_needing_layout()
-    }
-
-    /// Returns the nodes needing paint.
-    ///
-    /// These are repaint boundaries that need to be painted in the next
-    /// paint phase.
-    #[inline]
-    pub fn nodes_needing_paint(&self) -> &[DirtyNode] {
-        self.scheduler.nodes_needing_paint()
-    }
-
-    /// Returns the nodes needing compositing bits update.
-    #[inline]
-    pub fn nodes_needing_compositing_bits_update(&self) -> &[DirtyNode] {
-        self.scheduler.nodes_needing_compositing_bits_update()
-    }
-
-    /// Returns the nodes needing semantics update.
-    #[inline]
-    pub fn nodes_needing_semantics(&self) -> &[DirtyNode] {
-        self.scheduler.nodes_needing_semantics()
-    }
-
-    /// Adds a node to the layout dirty list.
-    ///
-    /// Routes into the mid-phase side queue when layout is active, otherwise
-    /// into `dirty.needs_layout`. See `DirtyTracker::add_node_needing_layout`
-    /// for the full routing and dedup contract.
-    pub fn add_node_needing_layout(&mut self, node_id: RenderId, depth: usize) {
-        self.scheduler.add_node_needing_layout(node_id, depth);
-    }
-
-    /// Marks a node as needing layout, propagating the `NEEDS_LAYOUT` flag
-    /// up the ancestor chain and pushing the **relayout boundary** onto
-    /// `dirty.needs_layout` for the next `run_layout` pass.
-    ///
-    /// **D-block PR-A1 U15** (memo D3) — ports Flutter's `markNeedsLayout`
-    /// walk (`.flutter/.../object.dart:2658-2700`). Thin forwarder: the walk
-    /// logic lives in `DirtyTracker::mark_needs_layout` so it is
-    /// unit-testable without a full owner. The `scheduler` and `render_tree`
-    /// fields are disjoint, so the split borrow compiles.
-    pub fn mark_needs_layout(&mut self, id: RenderId) {
-        self.scheduler.mark_needs_layout(&mut self.render_tree, id);
-    }
-
-    // ========================================================================
-    // Intrinsic / Dry-Layout Queries (memoized walks)
-    // ========================================================================
-
-    /// One intrinsic dimension of a box subtree, memoized per node.
-    ///
-    /// The walk mirrors Flutter's `getMinIntrinsicWidth`-family wrapper
-    /// layer (box.dart `_computeIntrinsics`): every node's answer for
-    /// `(dimension, extent)` is cached in its `RenderState` layout
-    /// cache, and `mark_needs_layout` clears the cache with
-    /// boundary-crossing escalation. Repeated probes of the same child
-    /// at the same extent — the canonical N-child container pattern —
-    /// cost one computation each.
-    ///
-    /// Average O(subtree) on a cold cache, O(1) per cached node;
-    /// worst case adds the hash-collision factor of the per-node maps.
-    ///
-    /// # Errors
-    ///
-    /// [`RenderError::NodeNotFound`](crate::error::RenderError::NodeNotFound)
-    /// for a stale/foreign id,
-    /// [`RenderError::ProtocolMismatch`](crate::error::RenderError::ProtocolMismatch)
-    /// if the subtree contains a sliver node (box intrinsics are
-    /// undefined there).
-    pub fn box_intrinsic_dimension(
-        &mut self,
-        id: RenderId,
-        dimension: crate::storage::IntrinsicDimension,
-        extent: f32,
-    ) -> crate::error::RenderResult<f32> {
-        #[cfg(any(test, feature = "testing"))]
-        let parent_data_seeds = self.parent_data_seeds.clone();
-        let mut slots = self.acquire_query_slots(id)?;
-        intrinsic_query(
-            &mut slots,
-            id,
-            dimension,
-            extent,
-            #[cfg(any(test, feature = "testing"))]
-            &parent_data_seeds,
-            #[cfg(not(any(test, feature = "testing")))]
-            &(),
-        )
-    }
-
-    /// The size a box subtree WOULD take under `constraints`, memoized
-    /// per `(node, constraints)` — Flutter's `getDryLayout`.
-    ///
-    /// # Errors
-    ///
-    /// Same surface as [`Self::box_intrinsic_dimension`].
-    pub fn box_dry_layout(
-        &mut self,
-        id: RenderId,
-        constraints: crate::constraints::BoxConstraints,
-    ) -> crate::error::RenderResult<flui_types::Size> {
-        let mut slots = self.acquire_query_slots(id)?;
-        dry_layout_query(&mut slots, id, constraints)
-    }
-
-    /// The dry baseline of a box node for `constraints`, memoized per
-    /// `(constraints, baseline)` — Flutter's `getDryBaseline`. The
-    /// computed answer may be `None` ("no baseline"); that answer is
-    /// cached too.
-    ///
-    /// # Errors
-    ///
-    /// Same surface as [`Self::box_intrinsic_dimension`].
-    pub fn box_dry_baseline(
-        &mut self,
-        id: RenderId,
-        constraints: crate::constraints::BoxConstraints,
-        baseline: crate::traits::TextBaseline,
-    ) -> crate::error::RenderResult<Option<f32>> {
-        let mut slots = self.acquire_query_slots(id)?;
-        dry_baseline_query(&mut slots, id, constraints, baseline)
-    }
-
-    /// Acquires the take-out borrow map for a memoizing query walk:
-    /// disjoint `&mut` over the subtree (the same `get_subtree_mut`
-    /// primitive the layout walk uses) plus each node's child-id
-    /// snapshot. A node is moved OUT of its slot while its own
-    /// computation runs, so re-entry — a child-link cycle — is
-    /// detectable instead of UB.
-    fn acquire_query_slots(
-        &mut self,
-        id: RenderId,
-    ) -> crate::error::RenderResult<rustc_hash::FxHashMap<RenderId, QuerySlot<'_>>> {
-        let ids = self.render_tree.collect_subtree_ids(id);
-        let nodes = self
-            .render_tree
-            .get_subtree_mut(&ids)
-            .ok_or(crate::error::RenderError::NodeNotFound(id))?;
-        Ok(ids
-            .iter()
-            .zip(nodes)
-            .map(|(&node_id, node)| {
-                let children = node.children().to_vec();
-                (
-                    node_id,
-                    QuerySlot {
-                        node: Some(node),
-                        children,
-                    },
-                )
-            })
-            .collect())
-    }
-
-    /// Adds a node to the paint dirty list.
-    ///
-    /// Routes into the mid-phase side queue when paint is active, otherwise
-    /// into `dirty.needs_paint`. See `DirtyTracker::add_node_needing_paint`
-    /// for the full routing and dedup contract.
-    pub fn add_node_needing_paint(&mut self, node_id: RenderId, depth: usize) {
-        self.scheduler.add_node_needing_paint(node_id, depth);
-    }
-
-    /// Adds a node to the compositing bits dirty list.
-    ///
-    /// Also sets `NEEDS_COMPOSITING_BITS_UPDATE` on the node (atomic) so the
-    /// `run_compositing` walk's short-circuit cannot drop this entry. Routes
-    /// via `debug_doing_layout` (compositing shares the layout flag). See
-    /// `DirtyTracker::add_node_needing_compositing_bits_update` for the
-    /// full contract.
-    ///
-    /// The `scheduler` and `render_tree` are disjoint fields: the shared
-    /// borrow `&self.render_tree` passed to the scheduler compiles cleanly.
-    pub fn add_node_needing_compositing_bits_update(&mut self, node_id: RenderId, depth: usize) {
-        self.scheduler
-            .add_node_needing_compositing_bits_update(&self.render_tree, node_id, depth);
-    }
-
-    /// Adds a node to the semantics dirty list.
-    ///
-    /// Routes via `debug_doing_semantics`. See
-    /// `DirtyTracker::add_node_needing_semantics` for the full contract.
-    pub fn add_node_needing_semantics(&mut self, node_id: RenderId, depth: usize) {
-        self.scheduler.add_node_needing_semantics(node_id, depth);
-    }
-
-    // ========================================================================
-    // Semantics enablement (data access, phase-agnostic)
-    // ========================================================================
-
-    /// Returns whether semantics are enabled.
-    #[inline]
-    pub fn semantics_enabled(&self) -> bool {
-        self.semantics_enabled.load(Ordering::Relaxed)
-    }
-
-    /// Sets whether semantics are enabled.
-    pub fn set_semantics_enabled(&self, enabled: bool) {
-        let was_enabled = self.semantics_enabled.swap(enabled, Ordering::Relaxed);
-        if enabled && !was_enabled {
-            self.notifier.read().fire_semantics_owner_created();
-        } else if !enabled && was_enabled {
-            self.notifier.read().fire_semantics_owner_disposed();
-        }
-    }
-
-    // ========================================================================
-    // Debug
-    // ========================================================================
-
-    /// Returns whether layout is currently being performed.
-    #[inline]
-    pub fn debug_doing_layout(&self) -> bool {
-        self.scheduler.debug_doing_layout()
-    }
-
-    /// Returns whether paint is currently being performed.
-    #[inline]
-    pub fn debug_doing_paint(&self) -> bool {
-        self.scheduler.debug_doing_paint()
-    }
-
-    /// Returns whether semantics update is currently being performed.
-    #[inline]
-    pub fn debug_doing_semantics(&self) -> bool {
-        self.scheduler.debug_doing_semantics()
-    }
-
-    /// Returns whether any pipeline phase is currently active.
-    #[inline]
-    pub fn debug_doing_any_phase(&self) -> bool {
-        self.scheduler.debug_doing_any_phase()
-    }
-
-    /// Returns the total number of dirty nodes across all lists.
-    #[inline]
-    pub fn dirty_node_count(&self) -> usize {
-        self.scheduler.dirty_node_count()
-    }
-
-    /// Returns whether there are any dirty nodes.
-    #[inline]
-    pub fn has_dirty_nodes(&self) -> bool {
-        self.scheduler.has_dirty_nodes()
-    }
-
-    /// Clears all dirty node lists without processing them.
-    ///
-    /// Use with caution — this discards pending work.
-    pub fn clear_all_dirty_nodes(&mut self) {
-        self.scheduler.clear_all();
-    }
-
-    /// Drains the mid-phase side queue into the active `dirty` set.
-    ///
-    /// Thin forwarder to `DirtyTracker::drain_mid_marks`. Available for
-    /// external callers (e.g. `flui-app`, integration tests) that need to
-    /// drain mid-marks between phase invocations without going through the
-    /// full phase-exit flow. The `run_*` phase methods drain via
-    /// `DirtyTracker::exit_phase` internally and do NOT call this method.
-    #[must_use]
-    pub fn drain_mid_layout_marks(&mut self) -> usize {
-        self.scheduler.drain_mid_marks()
-    }
-
-    /// Returns whether any mid-phase marks are pending drain.
-    #[inline]
-    pub fn has_mid_layout_marks(&self) -> bool {
-        self.scheduler.has_mid_marks()
-    }
-}
-
-// ============================================================================
-// Layout phase: run_layout + helpers
-// ============================================================================
-
-impl PipelineOwner<Layout> {
-    /// Transitions a layout-phase pipeline into the [`Compositing`] phase.
-    #[must_use]
-    pub fn into_compositing(self) -> PipelineOwner<Compositing> {
-        rebind_phase(self)
-    }
-
-    /// Returns to [`Idle`] from the layout phase (e.g. on error abort).
-    #[must_use]
-    pub fn into_idle(self) -> PipelineOwner<Idle> {
-        rebind_phase(self)
-    }
-
-    /// Updates layout for all dirty render objects.
-    ///
-    /// This is phase 1 of the rendering pipeline. During layout:
-    /// - Sizes and positions are calculated
-    /// - Objects may dirty paint or compositing
-    ///
-    /// Nodes are sorted by depth (shallow first) so parents are laid out
-    /// before their children. This matches Flutter's `flushLayout` behavior.
-    ///
-    /// # Synchronous Child Layout
-    ///
-    /// With interior mutability (RwLock on RenderNode), parent's
-    /// `perform_layout` can call `layout_child()` which triggers
-    /// synchronous child layout through the RenderTree. The child is laid
-    /// out immediately and returns its size.
-    pub fn run_layout(&mut self) -> crate::error::RenderResult<()> {
-        let _span =
-            tracing::debug_span!("layout", dirty_nodes = self.scheduler.layout_queue_len(),)
-                .entered();
-
-        // Process own dirty nodes if any
-        // Flutter pattern: while loop to handle nodes added during layout
-        while self.scheduler.has_layout_work() {
-            self.scheduler.enter_phase(PhaseKind::Layout);
-
-            // Take the dirty nodes and replace with empty set
-            // This allows new nodes to be added during layout (routed
-            // to mid_layout_marks per U22; drained back at end of
-            // iteration below).
-            let dirty_nodes = self.scheduler.take_layout_batch_shallow_first();
-
-            tracing::debug!(
-                "run_layout: sorted order (shallow-first) = {:?}",
-                dirty_nodes
-                    .iter()
-                    .map(|n| (n.id, n.depth))
-                    .collect::<Vec<_>>()
-            );
-
-            // Process each dirty node.
-            //
-            // **D-block PR-A1 U23:** layout_dirty_root replaces the
-            // legacy layout_node_with_children no-op recursion. Each
-            // dirty entry is laid out via the pre-acquired-subtree
-            // walk (U20.1) protected by LayoutCycleGuard (U21).
-            // Constraints come from cached state (post-frame-1) OR
-            // the binding-set root_constraints (frame-1 root).
-            for dirty_node in dirty_nodes {
-                // PR-A1 U23 P2 review fix (Copilot 3294417924): skip
-                // entries whose NEEDS_LAYOUT flag was already cleared
-                // earlier in this iteration. Common case: a parent's
-                // layout_child callback recursively lays out a child
-                // whose dirty-queue entry was queued separately
-                // (e.g., insert_child_render_object enqueues both).
-                // Re-laying out the child would be redundant +
-                // potentially side-effectful.
-                let already_clean = self
-                    .render_tree
-                    .get(dirty_node.id)
-                    .is_some_and(|n| !n.needs_layout());
-                if already_clean {
-                    tracing::trace!(
-                        id = ?dirty_node.id,
-                        "run_layout: skipping dirty-queue entry whose NEEDS_LAYOUT \
-                         was already cleared this iteration",
-                    );
-                    continue;
-                }
-
-                let Some(constraints) = self.cached_or_root_constraints(dirty_node.id) else {
-                    // PR-A1 U23 P2 review fix (Copilot 3294417942):
-                    // dropping the dirty entry here without recovery
-                    // strands the work. The two real cases this hits:
-                    //   1. Root id with root_constraints unset — the
-                    //      binding should have called
-                    //      set_root_constraints BEFORE run_frame.
-                    //      U23's set_root_constraints fix auto-marks
-                    //      root dirty when constraints land, so the
-                    //      next run_layout picks up the deferred
-                    //      work automatically.
-                    //   2. Non-root id with no cached constraints
-                    //      AND no parent-driven layout yet — the
-                    //      shallow-first dirty queue sort means the
-                    //      parent should have processed first; if
-                    //      it didn't (parent's perform_layout didn't
-                    //      call layout_child for this id), the entry
-                    //      is correctly dropped because the parent
-                    //      is the authority on child constraints.
-                    // Logged at warn so the diagnostic surfaces but
-                    // doesn't halt the pipeline.
-                    tracing::warn!(
-                        id = ?dirty_node.id,
-                        is_root = ?(self.root_id == Some(dirty_node.id)),
-                        "run_layout: no cached state.constraints() AND no \
-                         root_constraints (or id != root_id); skipping dirty entry. \
-                         Recovery: for root → call set_root_constraints (which \
-                         auto-marks the root dirty); for non-root → parent's \
-                         perform_layout must call ctx.layout_child(idx, c) for \
-                         this id first."
-                    );
-                    continue;
-                };
-                if let Err(e) = self.layout_dirty_root(dirty_node.id, constraints) {
-                    // PR-A1 U22 P1 review fix (Codex 3294365736): drain
-                    // mid-phase marks back into `dirty` even on error
-                    // path so they survive across phase invocations.
-                    let _ = self.scheduler.exit_phase(PhaseKind::Layout);
-                    return Err(e);
-                }
-
-                // Flutter parity (object.dart: `RenderObject.layout`
-                // unconditionally ends with `markNeedsPaint()`): a
-                // subtree that re-laid out must repaint, otherwise a
-                // pure-layout invalidation (setState moving a child)
-                // leaves stale pixels on screen. One entry per dirty
-                // root suffices — run_paint walks the whole tree from
-                // the root, so triggering the phase is what matters.
-                self.add_node_needing_paint(dirty_node.id, dirty_node.depth);
-            }
-
-            // PR-A1 U22 P1 review fix (Codex 3294365736): exit_phase clears
-            // debug_doing_layout AND drains mid_layout_marks back into `dirty`
-            // so the outer while condition picks up marks routed to the side
-            // queue during this iteration's `debug_doing_layout = true` window.
-            let _ = self.scheduler.exit_phase(PhaseKind::Layout);
-        }
-
-        // Drain deferred mutations: render objects may have enqueued
-        // child insertions, removals, or updates during layout. Apply
-        // them now, outside the `&mut` borrow scope of the layout walk.
-        // This is the Rust-native alternative to Flutter's
-        // `invokeLayoutCallback`.
-        //
-        // U3c D3 — true Remove → Insert → Update ordering. Previously the
-        // apply loop was strict FIFO with only a remove-vs-update skip,
-        // which let a frame that enqueues both Remove and Insert apply them
-        // in arbitrary order. Stable-partition into three buckets and apply
-        // each bucket in order. This makes the ordering comment above
-        // factually true rather than aspirational.
-        if !self.deferred_mutations.is_empty() {
-            let mutations = self.deferred_mutations.drain();
-
-            let (mut removes, rest): (Vec<_>, Vec<_>) = mutations
-                .into_iter()
-                .partition(|m| matches!(m, super::deferred::DeferredMutation::Remove { .. }));
-            let (inserts, updates): (Vec<_>, Vec<_>) = rest
-                .into_iter()
-                .partition(|m| matches!(m, super::deferred::DeferredMutation::Insert { .. }));
-
-            // Collect removed IDs for conflict detection (Update on same target).
-            let removed_ids: rustc_hash::FxHashSet<RenderId> = removes
-                .iter()
-                .filter_map(|m| match m {
-                    super::deferred::DeferredMutation::Remove { child_id, .. } => Some(*child_id),
-                    _ => None,
-                })
-                .collect();
-
-            tracing::trace!(
-                "run_layout: applying {} deferred mutations ({} removes, {} inserts, {} updates)",
-                removes.len() + inserts.len() + updates.len(),
-                removes.len(),
-                inserts.len(),
-                updates.len(),
-            );
-
-            let mut applied = 0usize;
-            let mut skipped = 0usize;
-
-            // Phase 1: Removes.
-            for mutation in removes.drain(..) {
-                self.apply_deferred_mutation(mutation);
-                applied += 1;
-            }
-
-            // Phase 2: Inserts.
-            for mutation in inserts {
-                self.apply_deferred_mutation(mutation);
-                applied += 1;
-            }
-
-            // Phase 3: Updates — skip any whose target was removed this batch.
-            for mutation in updates {
-                match &mutation {
-                    super::deferred::DeferredMutation::Update { target_id, .. }
-                        if removed_ids.contains(target_id) =>
-                    {
-                        tracing::warn!(
-                            ?target_id,
-                            "apply_deferred_mutation: skipping update — target was \
-                             removed in the same layout pass"
-                        );
-                        skipped += 1;
-                    }
-                    _ => {
-                        self.apply_deferred_mutation(mutation);
-                        applied += 1;
-                    }
-                }
-            }
-
-            if skipped > 0 {
-                tracing::warn!(
-                    "run_layout: {skipped} deferred mutations skipped due to \
-                     remove-update conflicts ({applied} applied)"
-                );
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Applies a single deferred mutation.
-    ///
-    /// Called after the layout pass completes. The mutation queue was
-    /// populated during layout by render objects that needed to modify
-    /// the tree structure (e.g., `LayoutBuilder`, `OverlayPortal`).
-    fn apply_deferred_mutation(&mut self, mutation: super::deferred::DeferredMutation) {
-        use super::deferred::{DeferredMutation, DeferredRenderObject};
-        match mutation {
-            DeferredMutation::Insert {
-                parent_id,
-                render_object,
-                index,
-                logical_index,
-                initial_parent_data,
-            } => {
-                // A deferred insert must schedule the new child (and re-dirty
-                // its parent) for layout and paint — mirroring
-                // `insert_child_render_object`. Without this the fresh node
-                // carries NEEDS_LAYOUT but is absent from every dirty queue,
-                // so it is laid out never and painted never (invisible child).
-                let Some(parent_depth) = self.render_tree.depth(parent_id) else {
-                    tracing::warn!(
-                        ?parent_id,
-                        "apply_deferred_mutation: Insert parent does not exist; mutation dropped"
-                    );
-                    return;
-                };
-                let child_depth = (parent_depth + 1) as usize;
-                let inserted = match render_object {
-                    DeferredRenderObject::Box(obj) => {
-                        self.render_tree.insert_box_child(parent_id, obj)
-                    }
-                    DeferredRenderObject::Sliver(obj) => {
-                        self.render_tree.insert_sliver_child(parent_id, obj)
-                    }
-                };
-                let Some(child_id) = inserted else { return };
-
-                // `insert_*_child` appends; honor an explicit position by
-                // moving the freshly appended child into place. Clamp so an
-                // out-of-range index lands at the end rather than panicking.
-                if let Some(i) = index
-                    && let Some(parent) = self.render_tree.get_mut(parent_id)
-                {
-                    parent.remove_child(child_id);
-                    let clamped = i.min(parent.child_count());
-                    parent.insert_child(clamped, child_id);
-                }
-
-                // D1 (lazy-sliver re-entrant build contract): install the
-                // logical index on the fresh child's parent-data so the lazy
-                // sliver consumer can reconcile it on the next pass.
-                //
-                // Fresh `RenderNode`s start with `parent_data = None`; the
-                // build backend seeds `initial_parent_data` with a pre-built
-                // `SliverMultiBoxAdaptorParentData { index: logical_index }`
-                // for exactly this case.  If parent-data is already present
-                // (e.g. a re-inserted node whose data survived) we stamp the
-                // index field directly instead so we never overwrite unrelated
-                // fields the existing box carries.
-                if (logical_index.is_some() || initial_parent_data.is_some())
-                    && let Some(child_node) = self.render_tree.get_mut(child_id)
-                {
-                    match child_node.parent_data_mut() {
-                        None => {
-                            // Node has no parent-data yet: install the
-                            // pre-built box wholesale.
-                            if let Some(pd) = initial_parent_data {
-                                child_node.set_parent_data(pd);
-                                tracing::trace!(
-                                    ?child_id,
-                                    logical_index,
-                                    "apply_deferred_mutation: installed \
-                                     initial parent-data on fresh child",
-                                );
-                            }
-                        }
-                        Some(pd) => {
-                            // Parent-data already present: stamp only the
-                            // logical-index field so other fields are
-                            // preserved.
-                            if let Some(li) = logical_index
-                                && let Some(lip) = pd.as_logical_index_mut()
-                            {
-                                lip.set_logical_index(li);
-                                tracing::trace!(
-                                    ?child_id,
-                                    logical_index = li,
-                                    "apply_deferred_mutation: stamped \
-                                     logical index into existing parent-data",
-                                );
-                            }
-                        }
-                    }
-                }
-
-                self.bootstrap_repaint_boundary_flag(child_id);
-                self.add_node_needing_layout(child_id, child_depth);
-                self.add_node_needing_paint(child_id, child_depth);
-                // `mark_needs_layout` (not `add_node_needing_layout`) so the
-                // parent's NEEDS_LAYOUT flag is actually set and its relayout
-                // boundary is enqueued — the dirty-root walk skips queued
-                // entries whose flag is clear, so a flagless enqueue would
-                // leave the new child unpositioned at the origin.
-                self.mark_needs_layout(parent_id);
-                tracing::trace!(
-                    ?parent_id,
-                    ?child_id,
-                    "apply_deferred_mutation: inserted child and scheduled layout + paint"
-                );
-            }
-            DeferredMutation::Remove {
-                parent_id,
-                child_id,
-            } => {
-                // Cascade-dispose: `remove_render_object` evicts the whole
-                // subtree from every dirty queue and frees it recursively
-                // (the old `remove_shallow` orphaned + leaked descendants and
-                // left stale dirty entries). Then re-dirty the parent so it
-                // reflows without the removed child.
-                let removed = self.remove_render_object(child_id);
-                if removed > 0 {
-                    // Re-dirty the parent (flag-setting walk, not a flagless
-                    // enqueue) so it reflows without the removed child.
-                    self.mark_needs_layout(parent_id);
-                }
-                tracing::trace!(
-                    ?parent_id,
-                    ?child_id,
-                    removed,
-                    "apply_deferred_mutation: removed subtree and re-dirtied parent"
-                );
-            }
-            DeferredMutation::Update { target_id, updater } => {
-                if let Some(node) = self.render_tree.get_mut(target_id) {
-                    match node {
-                        crate::storage::RenderNode::Box(entry) => {
-                            updater(entry.render_object_mut() as &mut dyn std::any::Any);
-                        }
-                        crate::storage::RenderNode::Sliver(entry) => {
-                            updater(entry.render_object_mut() as &mut dyn std::any::Any);
-                        }
-                    }
-                    tracing::trace!(?target_id, "apply_deferred_mutation: updated render object");
-                }
-            }
-        }
-    }
-
-    /// Returns the constraints to apply when laying out `id` as a
-    /// dirty root.
-    ///
-    /// **D-block PR-A1 U23:** sourced from (in priority order):
-    ///
-    /// 1. The node's cached `state.constraints()` — set on the
-    ///    previous frame's successful layout. This is the common
-    ///    case for re-layout (constraints unchanged → cache hit
-    ///    fast path inside `layout_dirty_root`).
-    /// 2. The binding-set [`Self::root_constraints`] if `id` is the
-    ///    tree root (`root_id`). Used on the very first frame
-    ///    before any layout has cached its own constraints.
-    /// 3. Otherwise `None` — caller skips this dirty entry with a
-    ///    warning (no constraints available means the parent's
-    ///    perform_layout must propagate constraints first; the
-    ///    dirty queue's shallow-first ordering ensures parents are
-    ///    processed before children).
-    fn cached_or_root_constraints(&self, id: RenderId) -> Option<BoxConstraints> {
-        // The binding-set root constraints are AUTHORITATIVE for the
-        // root: on resize, set_root_constraints updates them and marks
-        // the root dirty — if the stale cached constraints won here,
-        // the resize relayout would run at the OLD window size and the
-        // newly exposed area would stay unpainted.
-        if self.root_id == Some(id)
-            && let Some(root) = self.root_constraints
-        {
-            return Some(root);
-        }
-        if let Some(node) = self.render_tree.get(id)
-            && let Some(entry) = node.as_box()
-            && let Some(cached) = entry.state().constraints()
-        {
-            return Some(*cached);
-        }
-        if self.root_id == Some(id) {
-            return self.root_constraints;
-        }
-        None
-    }
-
-    /// D-block PR-A1b3 U20 — production disjoint-borrow layout walk.
-    ///
-    /// Lays out the subtree rooted at `id` with the supplied
-    /// `constraints`, running `RenderObject::perform_layout_raw` against a
-    /// typed [`crate::protocol::BoxLayoutCtx`] populated with the parent's direct children
-    /// (companion memo D1). Returns the parent's computed `Size` on
-    /// success.
-    ///
-    /// Replaces the recursion shape of `layout_node_with_children`
-    /// (which only walks the dirty tree without invoking per-node
-    /// layout — see the audit comment in that method). The
-    /// pipeline-side `run_layout` outer loop is rewired to this method
-    /// in U23.
-    ///
-    /// # Mechanism (U20.1 — pre-acquired subtree borrows)
-    ///
-    /// 1. **Collect ids**: `RenderTree::collect_subtree_ids(id)` walks
-    ///    the subtree in DFS pre-order producing
-    ///    `Vec<RenderId>` covering root + all descendants.
-    /// 2. **Pre-acquire borrows in ONE scope**:
-    ///    `RenderTree::get_subtree_mut(&ids)` materialises N disjoint
-    ///    `&mut RenderNode` references in a single function body via
-    ///    the proven `*mut Slab` reborrow pattern that already powers
-    ///    [`RenderTree::get_two_mut`] and
-    ///    [`RenderTree::get_parent_and_children_mut`].
-    /// 3. **Index by id**: the N borrows are wrapped in a private
-    ///    `SubtreeArena` as a `HashMap<RenderId, (NodePtr, AtomicBool)>`
-    ///    (raw pointer alias of the still-live `&mut RenderNode` borrows,
-    ///    paired with a per-slot in-flight flag). Lookup is O(1) by id.
-    /// 4. **Recursive walk**: a private `layout_subtree_borrowed`
-    ///    helper indexes into `SubtreeArena` to acquire one node's
-    ///    reborrow at each call level. The leaf path delegates to
-    ///    [`RenderEntry::layout_leaf_only`](crate::storage::RenderEntry::layout_leaf_only).
-    ///    The non-leaf path constructs a Direct-storage `BoxLayoutCtx`
-    ///    via the erased driver context ([`crate::protocol::ErasedBoxLayoutCtx`]) with a closure
-    ///    that captures `&SubtreeArena` (Sync via `NodePtr`'s
-    ///    `unsafe impl`) and re-enters `layout_subtree_borrowed` for
-    ///    each child. The bridge in `traits/render_box.rs` reconstructs
-    ///    a typed `BoxLayoutCtx<T::Arity, T::ParentData>` (Proxy
-    ///    variant) and forwards to `RenderBox::perform_layout`.
-    ///    Synchronous `ctx.layout_child(i, c)` calls dispatch through
-    ///    the callback, recursing into the child's subtree via its
-    ///    pre-acquired `NodePtr`.
-    /// 5. **Per-level cleanup**: on success **AND when no descendant
-    ///    errored**, updates `state.set_geometry` /
-    ///    `state.set_constraints`, bootstraps `IS_RELAYOUT_BOUNDARY`,
-    ///    clears `NEEDS_LAYOUT`. When a descendant errored
-    ///    mid-callback, geometry + constraints are still recorded but
-    ///    `NEEDS_LAYOUT` stays set on the parent so the next dirty
-    ///    walk re-runs the subtree.
-    ///
-    /// # Soundness (U20.1 fix — Miri-clean)
-    ///
-    /// The prior PR-A1b3 design (PR #144) used a recursive raw-pointer
-    /// re-entry into `RenderTree` from inside the layout-child callback —
-    /// outer `&mut RenderEntry` was held LIVE across `perform_layout_raw`
-    /// while the inner call synthesised a fresh `&mut RenderTree` from
-    /// the same `*mut`. Under Stacked / Tree Borrows that invalidates
-    /// the outer tag (latent UB; Miri flagged the 2-level and 3-level
-    /// happy paths).
-    ///
-    /// The U20.1 redesign eliminates the inner reborrow entirely. All
-    /// N disjoint `&mut RenderNode` borrows are acquired in ONE call to
-    /// `get_subtree_mut` (single `&mut Slab` reborrow scope, mirroring
-    /// `get_parent_and_children_mut`). The callback then acquires one
-    /// node's reborrow per call level by dereferencing the pre-acquired
-    /// `NodePtr` for that slot — no `&mut RenderTree` ever appears
-    /// inside the callback chain.
-    ///
-    /// At any given moment during the walk, the per-slot reborrows in
-    /// scope are: one for the current call's parent + one for each
-    /// active recursive child call below it. All on **distinct slab
-    /// slots** (parent ≠ child in a well-formed tree). The Unique tag
-    /// for each slot lives independently because `NodePtr` is a raw
-    /// pointer (SharedReadWrite permission on the allocation, distinct
-    /// derived Unique tags per slot reborrow). Miri verifies this on
-    /// the existing U20 integration tests.
-    ///
-    /// # Error handling
-    ///
-    /// - **Leaf-path panics** in user `perform_layout` → caught by
-    ///   `layout_leaf_only`'s `catch_unwind`, returned as
-    ///   [`crate::error::RenderError::Poisoned`].
-    /// - **Non-leaf-path panics** (PR-A1b3 review fix): wrapped in
-    ///   `catch_unwind` at the non-leaf `perform_layout_raw` call site,
-    ///   returned as the same [`crate::error::RenderError::Poisoned`].
-    ///   Symmetric with the leaf path.
-    /// - **Descendant `Err` returned through the callback** → tracking
-    ///   flag (`AtomicBool`) set; outer `perform_layout` still completes
-    ///   with `Size::ZERO` for that child; outer `Ok` is returned to the
-    ///   caller, BUT parent's `NEEDS_LAYOUT` is **not cleared**.
-    ///   Next-frame dirty walk re-runs the parent. The `LayoutChildCallback`
-    ///   signature is `Fn(_) -> Size`, not `Fn(_) -> Result<Size, _>`,
-    ///   so callback failures can't propagate as typed `Err` through
-    ///   the parent's `perform_layout` body. Surfacing the typed error
-    ///   to the outer caller requires widening `LayoutChildCallback`
-    ///   to `Result`; deferred to Core.1.
-    /// - **Stale tree state** (id not in tree → `NodeNotFound`; id is
-    ///   present but wrong protocol → `ProtocolMismatch`).
-    ///
-    /// # ParentData scope (current limitation)
-    ///
-    /// The pipeline-side Direct `BoxLayoutCtx` is parameterised over
-    /// [`crate::parent_data::BoxParentData`], so widgets whose `T::ParentData` is the default
-    /// (`RenderPadding`, `RenderCenter`, `RenderColoredBox`,
-    /// `RenderOpacity`, `RenderTransform`, `RenderSizedBox`) drive
-    /// through correctly. Non-default parent-data types (e.g.,
-    /// `FlexParentData` on `RenderFlex`) trigger a
-    /// `BoxLayoutCtx::from_erased` debug-assert mismatch in debug builds
-    /// and silently fail to downcast in release builds. Pipeline-driven
-    /// flex layout is out of scope for D-block; per-render-object
-    /// `T::ParentData` dispatch lands as a Core.1 follow-up alongside the
-    /// real `RenderFlex` slice integration.
-    ///
-    /// # Cycle / depth safety (U21 wired)
-    ///
-    /// Three-layer cycle protection (U20.1 + U21 combined):
-    ///
-    /// 1. `collect_subtree_ids` terminates safely on cycles via its
-    ///    `visited` `HashSet<RenderId>` short-circuit (PR #145) —
-    ///    the cyclic id is visited at most once, the cycle edge is
-    ///    silently dropped from the collected subtree, deduplicated
-    ///    `Vec<RenderId>` returns. No hang / OOM at the collect
-    ///    phase.
-    /// 2. `get_subtree_mut` receives the deduplicated id list →
-    ///    uniqueness precondition satisfied → returns `Some(refs)`.
-    ///    No double-borrow attempt at acquisition.
-    /// 3. `layout_subtree_borrowed` marks each `id`'s in-flight flag in
-    ///    `SubtreeArena::by_id` via the
-    ///    `LayoutCycleGuard` RAII on entry (U21). A `perform_layout`
-    ///    body that calls `layout_child` for an ancestor id already
-    ///    in flight hits the guard's `enter` collision check →
-    ///    returns [`crate::error::RenderError::LayoutCycle`]
-    ///    immediately instead of attempting a second `NodePtr`
-    ///    reborrow (which would be UB).
-    ///
-    /// The cycle error collapses through the layout-child callback
-    /// (Size::ZERO + `descendant_error_flag`) so the parent stays
-    /// `NEEDS_LAYOUT` for next-frame retry. The cycle persists
-    /// structurally so retry will re-surface `LayoutCycle` — but
-    /// predictably, never as panic/UB/hang. The user can fix the
-    /// tree (remove the cyclic `add_child`) and the next retry
-    /// succeeds.
-    ///
-    /// Frame-cross panic safety: `LayoutCycleGuard::Drop` runs on
-    /// every exit path including unwind from a panicking
-    /// `perform_layout`. Combined with the non-leaf path's
-    /// `catch_unwind` wrapper, the cycle set stays consistent across
-    /// frames — a panic does not leak an in-flight id.
-    ///
-    /// **U23 wiring is now soundness-unblocked.** `run_layout` may
-    /// wire `layout_dirty_root` per its dirty-queue iteration in U23.
-    pub fn layout_dirty_root(
-        &mut self,
-        id: RenderId,
-        constraints: BoxConstraints,
-    ) -> crate::error::RenderResult<Size> {
-        // Steps 1–3: collect subtree ids, pre-acquire disjoint &mut borrows,
-        // and wrap them in a SubtreeArena for O(1) by-id lookup during the
-        // recursive walk.  The unsafe aliasing machinery lives entirely inside
-        // `subtree_arena::SubtreeArena`; this call site is safe.
-        let arena = SubtreeArena::from_tree(
-            &mut self.render_tree,
-            id,
-            #[cfg(any(test, feature = "testing"))]
-            &self.parent_data_seeds,
-        )?;
-
-        // Step 4: recursive walk via the safe arena API.  Internally the arena
-        // reborrows one NodePtr per call level (distinct slots; parent ≠ child
-        // enforced by LayoutCycleGuard), but no `unsafe` appears here.
-        let result = arena.layout_child(id, constraints);
-
-        // Step 5: drain on-demand child builds AND removals the walk recorded
-        // (re-entrant build contract v1, ADR-0003 Decision 2 + U3c D2).
-        // Take both (owned), then DROP `arena` to release the &mut RenderTree
-        // subtree borrow before touching `&mut self`.
-        //
-        // D3 ordering: Remove → Insert → Update.  Removes are drained first so
-        // that an off-band child evicted this pass does not collide with the
-        // Insert that replaces it in the same batch.
-        let pending_removes = arena.take_pending_removes();
-        let pending_builds = arena.take_pending_builds();
-        drop(arena);
-
-        // Apply removes first (D3 ordering).  Each entry is `(parent, child)`:
-        // the parent is the sliver's own node_id (tagged at push time in
-        // ErasedSliverLayoutCtx::dispose_box_child), NOT the walk root `id`.
-        // Using `id` here would misdirect `mark_needs_layout` to the viewport
-        // root instead of the lazy sliver, preventing it from reflowing.
-        for (parent, child_id) in pending_removes {
-            self.defer_remove(parent, child_id);
-        }
-
-        for pending in pending_builds {
-            self.defer_insert_box(
-                pending.parent,
-                pending.object,
-                Some(pending.index),
-                Some(pending.logical_index),
-                pending.initial_parent_data,
-            );
-        }
-
-        result
-    }
-}
-
-// ============================================================================
-// Compositing phase: run_compositing
-// ============================================================================
-
-impl PipelineOwner<Compositing> {
-    /// Transitions a compositing-phase pipeline into the [`PaintPhase`] phase.
-    #[must_use]
-    pub fn into_paint(self) -> PipelineOwner<PaintPhase> {
-        rebind_phase(self)
-    }
-
-    /// Returns to [`Idle`] from the compositing phase.
-    #[must_use]
-    pub fn into_idle(self) -> PipelineOwner<Idle> {
-        rebind_phase(self)
-    }
-
-    /// Updates compositing bits for all dirty render objects.
-    ///
-    /// **D-block PR-A2 U34 (memo D3-3).** Port of Flutter's
-    /// `PipelineOwner.flushCompositingBits` + per-object
-    /// `RenderObject._updateCompositingBits`
-    /// (`.flutter/.../object.dart:3226-3258`). For each entry in
-    /// `dirty.needs_compositing` (sorted shallow-first to match
-    /// Flutter's `_nodesNeedingCompositingBitsUpdate.sort`), this
-    /// method recursively walks the subtree, recomputing
-    /// `NEEDS_COMPOSITING` bottom-up:
-    ///
-    /// 1. If a node does NOT have `NEEDS_COMPOSITING_BITS_UPDATE`
-    ///    set, skip (parent walk already covered, or no work to do).
-    /// 2. Otherwise: save `old_needs_compositing`, clear current
-    ///    `NEEDS_COMPOSITING`, recurse into each child, OR-in each
-    ///    child's `NEEDS_COMPOSITING` result.
-    /// 3. Force `NEEDS_COMPOSITING = true` if the node is a repaint
-    ///    boundary (`IS_REPAINT_BOUNDARY`) or always needs compositing
-    ///    (`RenderObject::always_needs_compositing()`).
-    /// 4. Three transition cases:
-    ///    - **Lost-boundary**: if the node previously was a repaint
-    ///      boundary (`WAS_REPAINT_BOUNDARY`) but no longer is, clear
-    ///      its accumulated paint state and re-enqueue for paint so
-    ///      a new boundary owner picks it up (Flutter object.dart:3246).
-    ///    - **Compositing changed**: if `old_needs_compositing !=
-    ///      new_needs_compositing`, mark dirty for paint so the
-    ///      compositor sees the new shape (Flutter object.dart:3252).
-    ///    - **No change**: clear `NEEDS_COMPOSITING_BITS_UPDATE` and
-    ///      leave paint state untouched (Flutter object.dart:3255).
-    ///
-    /// The walk is staged via a private `CompositingWalkActions`
-    /// accumulator so that post-walk paint-queue mutations don't
-    /// fight the recursive borrows: the recursion reads
-    /// `&self.render_tree` (shared) and accumulates actions, then
-    /// we apply them under `&mut self`.
-    pub fn run_compositing(&mut self) -> crate::error::RenderResult<()> {
-        // Empty fast-path: no allocation, no logging churn for the
-        // common "nothing changed" frame.
-        if !self.scheduler.has_compositing_work() {
-            return Ok(());
-        }
-        let _span = tracing::debug_span!(
-            "compositing",
-            dirty_nodes = self.scheduler.compositing_queue_len(),
-        )
-        .entered();
-
-        // Sort shallow-first per Flutter
-        // `_nodesNeedingCompositingBitsUpdate.sort((a, b) => a.depth - b.depth)`.
-        self.scheduler.sort_compositing_shallow_first();
-
-        // Iterate the dirty list by shared reference: the recursion takes
-        // `&self`, which coexists with the slice's shared borrow of
-        // the compositing queue (both shared). Queue mutations are deferred
-        // into `actions` and applied after the walk, so nothing mutates the
-        // list mid-iteration.
-        let mut actions = CompositingWalkActions::default();
-        let compositing_slice = self.scheduler.compositing_queue_as_slice();
-        for node in compositing_slice {
-            self.update_subtree_compositing_bits(node.id, &mut actions);
-        }
-
-        // Apply paint-queue mutations after the walk completes (under
-        // disjoint `&mut self`). Remove-first, then re-enqueue, so an
-        // id present in both buckets ends up correctly re-queued at the
-        // post-walk depth.
-        if !actions.remove_from_paint_queue.is_empty() {
-            self.scheduler
-                .retain_paint_queue(&actions.remove_from_paint_queue);
-        }
-        for (id, depth) in actions.mark_needs_paint {
-            self.add_node_needing_paint(id, depth);
-        }
-
-        // PR #109 retain-capacity idiom (kept from pre-U34 stub): `clear()`
-        // preserves the Vec's backing capacity across frames.
-        self.scheduler.clear_compositing_queue();
-        Ok(())
-    }
-
-    /// Recursive helper for [`Self::run_compositing`] — see that method's
-    /// doc for the algorithm. Operates on shared `&self.render_tree` via
-    /// interior-mutability flag accessors; paint-queue side effects are
-    /// staged via `actions` for post-walk application.
-    fn update_subtree_compositing_bits(&self, id: RenderId, actions: &mut CompositingWalkActions) {
-        ensure_stack(|| self.update_subtree_compositing_bits_impl(id, actions));
-    }
-
-    /// Body of [`Self::update_subtree_compositing_bits`]; split out so
-    /// every recursion level enters through the [`ensure_stack`] probe.
-    /// (Its frames are smaller than the layout walk's, so it survived
-    /// deeper trees by luck — same crash class, just a later threshold.)
-    fn update_subtree_compositing_bits_impl(
-        &self,
-        id: RenderId,
-        actions: &mut CompositingWalkActions,
-    ) {
-        let Some(node) = self.render_tree.get(id) else {
-            return;
-        };
-        if !node.needs_compositing_bits_update() {
-            return;
-        }
-
-        let old_needs_compositing = node.needs_compositing();
-        node.clear_needs_compositing();
-
-        // Iterate the child slice in-place — both `tree.children(id)`
-        // and the recursive `update_subtree_compositing_bits` call are
-        // shared borrows of `&self.render_tree`, so they coexist
-        // without a clone. Pre-fix the loop cloned children into a
-        // fresh `Vec<RenderId>` per visited node (per-node heap
-        // allocation) — flagged by Copilot #3294557204 as conflicting
-        // with the repo's documented "no per-node child clone"
-        // optimization in `RenderTree::visit_depth_first`
-        // (`storage/tree.rs:738-751`).
-        //
-        // Index loop (not iterator) so the loop body can call `&self`
-        // recursion without holding the slice iterator across the
-        // call (slice iter would borrow `&self.render_tree.children(id)`,
-        // which transitively borrows `&self`).
-        let child_count = self.render_tree.children(id).len();
-        for i in 0..child_count {
-            let child_id = self.render_tree.children(id)[i];
-            self.update_subtree_compositing_bits(child_id, actions);
-            if let Some(child) = self.render_tree.get(child_id)
-                && child.needs_compositing()
-            {
-                node.mark_needs_compositing();
-            }
-        }
-
-        if node.is_repaint_boundary_flag() || node.always_needs_compositing() {
-            node.mark_needs_compositing();
-        }
-
-        let new_needs_compositing = node.needs_compositing();
-        let is_boundary = node.is_repaint_boundary_flag();
-        let was_boundary = node.was_repaint_boundary();
-
-        // Flutter object.dart:3246 — lost-boundary status: drop the
-        // accumulated paint state so a NEW boundary parent picks this
-        // node up for paint. The id is removed from the dirty paint
-        // queue (since the queued paint targeted us-as-a-boundary)
-        // and re-enqueued at our current depth so the walk in
-        // `mark_needs_paint`'s spirit re-locates the responsible
-        // boundary owner.
-        if !is_boundary && was_boundary {
-            node.clear_needs_paint();
-            actions.remove_from_paint_queue.insert(id);
-            node.clear_needs_compositing_bits_update();
-            let depth = self.render_tree.depth(id).unwrap_or(0) as usize;
-            actions.mark_needs_paint.push((id, depth));
-        } else if old_needs_compositing != new_needs_compositing {
-            // Flutter object.dart:3252 — compositing shape changed:
-            // mark paint dirty so the compositor sees the new shape.
-            node.clear_needs_compositing_bits_update();
-            let depth = self.render_tree.depth(id).unwrap_or(0) as usize;
-            actions.mark_needs_paint.push((id, depth));
-        } else {
-            // Flutter object.dart:3255 — no shape change: just clear
-            // the bits-update flag.
-            node.clear_needs_compositing_bits_update();
-        }
-    }
-}
-
-/// Side-effects staged during a compositing-bits walk and applied
-/// after the recursion under `&mut self` (D-block PR-A2 U34).
-///
-/// The recursive walk in
-/// [`PipelineOwner::update_subtree_compositing_bits`] runs under
-/// `&self` (interior-mutability flag access only). Paint-queue
-/// mutations (remove-from / push-to `dirty.needs_paint`) can't
-/// happen mid-recursion without re-borrowing `&mut self`, so they
-/// are recorded here and replayed by [`PipelineOwner::run_compositing`]
-/// post-walk.
-#[derive(Default)]
-struct CompositingWalkActions {
-    /// `(id, depth)` pairs to enqueue via `add_node_needing_paint`
-    /// after the walk. Either the lost-boundary or compositing-shape-
-    /// changed branch may push here.
-    mark_needs_paint: Vec<(RenderId, usize)>,
-    /// Ids to drop from `dirty.needs_paint` before the re-enqueue
-    /// (lost-boundary branch only — a queued paint targeted the node
-    /// as-a-boundary, which it no longer is).
-    remove_from_paint_queue: rustc_hash::FxHashSet<RenderId>,
-}
-
-// ============================================================================
-// Paint phase: run_paint + helpers
-// ============================================================================
-
-impl PipelineOwner<PaintPhase> {
-    /// Transitions a paint-phase pipeline into the [`Semantics`] phase.
-    #[must_use]
-    pub fn into_semantics(self) -> PipelineOwner<Semantics> {
-        rebind_phase(self)
-    }
-
-    /// Returns to [`Idle`] from the paint phase.
-    #[must_use]
-    pub fn into_idle(self) -> PipelineOwner<Idle> {
-        rebind_phase(self)
-    }
-
-    /// Paints all dirty render objects.
-    ///
-    /// Phase 3 of the rendering pipeline, as a **fragment composition**
-    /// (sans-IO paint model): each node's `paint_raw` records a
-    /// node-local fragment — draw runs, child markers, clip scopes —
-    /// which is immediately replayed into the frame's [`LayerTree`].
-    /// Adjacent inline draw runs merge into shared `PictureLayer`s;
-    /// repaint-boundary children are rebased to `Offset::ZERO` under
-    /// their own `OffsetLayer`; clip scopes become real clip layers.
-    ///
-    /// A fresh full `LayerTree` is produced every paint pass —
-    /// cross-frame retention of boundary subtrees is deliberately out
-    /// of scope until the layer tree grows a structural-sharing
-    /// substrate and the engine an incremental upload path.
-    pub fn run_paint(&mut self) -> crate::error::RenderResult<()> {
-        if !self.scheduler.has_paint_work() {
-            return Ok(());
-        }
-
-        let _span = tracing::debug_span!("paint", dirty_nodes = self.scheduler.paint_queue_len(),)
-            .entered();
-
-        self.scheduler.enter_phase(PhaseKind::Paint);
-
-        // Deepest-first ordering retained (Flutter `flushPaint`): the
-        // full-tree descent below repaints everything, but per-boundary
-        // dirty-driven repaints will rely on this order once retention
-        // lands, and keeping it now means the dirty-list semantics
-        // don't shift under that change.
-        self.scheduler.sort_paint_deep_first();
-
-        if let Some(root_id) = self.root_id
-            && self.render_tree.get(root_id).is_some()
-        {
-            // Build a set of dirty node IDs for O(1) lookup during the
-            // paint walk.
-            let dirty_ids: rustc_hash::FxHashSet<RenderId> = self
-                .scheduler
-                .nodes_needing_paint()
-                .iter()
-                .map(|d| d.id)
-                .collect();
-
-            let mut composer = FragmentComposer::new(self.device_pixel_ratio);
-            match self.paint_subtree(&mut composer, root_id, Offset::ZERO, &dirty_ids) {
-                Ok(()) => {
-                    let layer_tree = composer.finish();
-                    tracing::debug!("run_paint: layer tree has {} layers", layer_tree.len());
-                    self.last_layer_tree = Some(layer_tree);
-                }
-                Err(e) => {
-                    // Restore the debug invariant before propagating so
-                    // the owner stays consistent on the error path.
-                    let _ = self.scheduler.exit_phase(PhaseKind::Paint);
-                    return Err(e);
-                }
-            }
-        }
-
-        // Dirty-list residue scan: any node still flagged needs_paint
-        // AFTER the root descent was not reached by it (multi-root or
-        // detached subtree). Warn + clear so the bug is visible AND the
-        // dirty list doesn't accumulate across frames.
-        for dirty_node in self.scheduler.nodes_needing_paint() {
-            if let Some(render_node) = self.render_tree.get(dirty_node.id)
-                && render_node.needs_paint()
-            {
-                tracing::warn!(
-                    id = ?dirty_node.id,
-                    depth = dirty_node.depth,
-                    "run_paint: dirty node not reached by root descent (multi-root \
-                     or detached subtree?); paint dropped, flag cleared"
-                );
-                render_node.clear_needs_paint();
-            }
-        }
-        // `clear()` retains capacity (preserve Vec backing across frames).
-        self.scheduler.clear_paint_queue();
-
-        // exit_phase clears debug_doing_paint AND drains mid-paint marks back
-        // into dirty so paint marks made during this pass become next-frame
-        // work rather than being stranded — Flutter's flushPaint semantics.
-        //
-        // Finding 2 (intentional improvement over pre-refactor behavior):
-        // exit_phase also drains mid-marks on the ERROR path (the early-return
-        // above calls exit_phase before returning Err). Pre-refactor, the error
-        // path only cleared debug_doing_paint and did NOT drain mid-marks, so
-        // any mark made between enter_phase and the error was silently lost.
-        // The always-drain contract of exit_phase is the correct behavior:
-        // mid-paint marks scheduled before the error survive into the next
-        // frame's retry rather than being dropped.
-        let _ = self.scheduler.exit_phase(PhaseKind::Paint);
-
-        Ok(())
-    }
-
-    /// Records one node's paint fragment and replays it into the
-    /// composer, recursing at child markers.
-    ///
-    /// Per-node order follows Flutter's `PaintingContext._paintWithContext`:
-    /// `WAS_REPAINT_BOUNDARY` is written and `NEEDS_PAINT` cleared
-    /// **before** the node paints, so a paint body that re-marks its own
-    /// node is caught by the debug check below instead of silently
-    /// erasing the evidence.
-    fn paint_subtree(
-        &self,
-        composer: &mut FragmentComposer,
-        node_id: RenderId,
-        origin: Offset,
-        dirty_set: &rustc_hash::FxHashSet<RenderId>,
-    ) -> crate::error::RenderResult<()> {
-        ensure_stack(|| self.paint_subtree_impl(composer, node_id, origin, dirty_set))
-    }
-
-    /// Body of [`Self::paint_subtree`]; split out so every recursion
-    /// level enters through the [`ensure_stack`] probe.
-    fn paint_subtree_impl(
-        &self,
-        composer: &mut FragmentComposer,
-        node_id: RenderId,
-        origin: Offset,
-        dirty_set: &rustc_hash::FxHashSet<RenderId>,
-    ) -> crate::error::RenderResult<()> {
-        let Some(render_node) = self.render_tree.get(node_id) else {
-            return Ok(());
-        };
-
-        let is_repaint_boundary = render_node.is_repaint_boundary();
-
-        let alpha = render_node.paint_alpha();
-        let layer_blend = render_node.paint_layer_blend();
-        let transform = render_node.paint_transform();
-        let child_ids: Vec<RenderId> = render_node.children().to_vec();
-
-        // Written unconditionally PRE-paint (Flutter object.dart:3560):
-        // a node flipping boundary→non-boundary leaves exactly one
-        // `WAS_REPAINT_BOUNDARY=true` trail for the next compositing
-        // walk's lost-boundary branch.
-        render_node.set_was_repaint_boundary(is_repaint_boundary);
-
-        // Clear BEFORE paint so the post-paint check catches a paint
-        // body that marks its own node dirty (paint-must-not-redirty).
-        render_node.clear_needs_paint();
-
-        // Fully transparent subtree: skip recording entirely. Children
-        // keep whatever dirty flags they carry; the residue scan in
-        // run_paint clears them with a warning.
-        // Uses `skip_paint()` rather than `alpha == Some(0)` so that
-        // `paint_alpha()` encoding only controls layer-emission; the
-        // skip-paint decision is a separate, explicit contract
-        // (Flutter: `if (_alpha == 0) return;` in RenderOpacity.paint).
-        if render_node.skip_paint() {
-            return Ok(());
-        }
-
-        // Flutter object.dart:3497 — a node that still needs layout must
-        // not paint stale geometry. Layout runs before paint in the
-        // pipeline, so this guards descendant-error and partial-frame
-        // paths where a poisoned layout left the flag set.
-        if render_node.needs_layout() {
-            return Ok(());
-        }
-
-        // Sliver visibility cull: a sliver with zero paint extent
-        // (`!visible`) paints nothing and splices no children (Flutter:
-        // the viewport skips invisible slivers). The gate lives here, in
-        // the driver — next to the sliver hit-test extent gate — so sliver
-        // objects no longer cache `geometry` just to short-circuit their
-        // own `paint`. Box nodes (`geometry_sliver() == None`) are never
-        // culled here. Same dirty-residue handling as the `alpha == 0`
-        // skip above.
-        if render_node.geometry_sliver().is_some_and(|g| !g.visible) {
-            return Ok(());
-        }
-
-        // Record the node's fragment. paint_raw sees ONLY the recorder
-        // (sans-IO): no tree access, no layer access, no recursion.
-        let debug_name = render_node.debug_name();
-        let mut recorder = FragmentRecorder::new(origin, self.device_pixel_ratio);
-        std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            render_node.paint_raw(&mut recorder, child_ids.len());
-        }))
-        .map_err(|_| crate::error::RenderError::poisoned(debug_name, "paint"))?;
-        let fragment = recorder.finish();
-
-        debug_assert!(
-            !render_node.needs_paint(),
-            "paint-must-not-redirty: a render object marked ITSELF \
-             needs-paint during its own paint; derive visual changes \
-             from state read at paint time instead of re-marking",
-        );
-
-        // Effect hooks wrap the ENTIRE node fragment (self draws AND
-        // children). The pre-fragment walk wrapped children only; hook
-        // implementors draw nothing themselves, so the visible result
-        // is identical and the new rule matches Flutter (RenderOpacity
-        // wraps its child's whole paint).
-        // Alpha–blend coupling: the layer is emitted only when `paint_alpha()` returns
-        // `Some`.  A render object that overrides `paint_layer_blend() -> Some(mode)`
-        // but leaves `paint_alpha() -> None` will silently drop the blend layer —
-        // the advanced compositor never sees it.  When wiring the first render-tree
-        // consumer of an advanced blend, override BOTH hooks and return `Some(255)`
-        // from `paint_alpha()` for an opaque-blend-only layer.
-        let mut effect_layers = 0usize;
-        if let Some(alpha) = alpha {
-            let alpha_f32 = f32::from(alpha) / 255.0;
-            let opacity_layer = match layer_blend {
-                Some(blend) => OpacityLayer::with_blend(alpha_f32, Offset::ZERO, blend),
-                None => OpacityLayer::with_offset(alpha_f32, Offset::ZERO),
-            };
-            composer.push_layer(Layer::Opacity(opacity_layer));
-            effect_layers += 1;
-        }
-        if let Some(matrix) = transform {
-            // The node reports its transform in LOCAL coordinates, but
-            // every run inside this layer space is recorded with the
-            // accumulated `origin` baked into its canvas transform.
-            // Conjugate by the origin (Flutter object.dart
-            // `pushTransform`: T(offset)·M·T(−offset)) so the matrix
-            // pivots around the node's own origin instead of the layer
-            // origin — a raw local matrix would translate/rotate the
-            // whole accumulated space.
-            let effective = if origin == Offset::ZERO {
-                matrix
-            } else {
-                let (dx, dy) = (origin.dx.get(), origin.dy.get());
-                flui_types::Matrix4::translation(dx, dy, 0.0)
-                    * matrix
-                    * flui_types::Matrix4::translation(-dx, -dy, 0.0)
-            };
-            composer.push_layer(Layer::Transform(TransformLayer::new(effective)));
-            effect_layers += 1;
-        }
-
-        for op in fragment.ops {
-            match op {
-                FragmentOp::Run(list) => composer.append_run(list),
-                FragmentOp::Push(clip) => composer.push_layer(clip_layer(*clip, origin)),
-                FragmentOp::Pop => composer.pop_layer(),
-                FragmentOp::Child {
-                    index,
-                    offset_override,
-                } => {
-                    let Some(&child_id) = child_ids.get(index) else {
-                        debug_assert!(
-                            false,
-                            "fragment child marker {index} out of range ({} children) — \
-                             PaintCx bounds-checks markers, so a mismatch means the \
-                             tree changed during paint",
-                            child_ids.len(),
-                        );
-                        continue;
-                    };
-                    let Some(child_node) = self.render_tree.get(child_id) else {
-                        continue;
-                    };
-                    if child_node
-                        .as_sliver()
-                        .and_then(|entry| entry.state().geometry())
-                        .is_some_and(|geometry| !geometry.visible)
-                    {
-                        continue;
-                    }
-                    // Authoritative child position: RenderState.offset,
-                    // committed by the layout walk; paint_child_at
-                    // overrides it explicitly.
-                    let child_offset = offset_override.unwrap_or_else(|| child_node.offset());
-                    let child_is_boundary = child_node.is_repaint_boundary();
-
-                    if child_is_boundary {
-                        // Boundary children rebase to ZERO under their
-                        // own OffsetLayer so a future offset-only move
-                        // is a layer-property update, not a repaint.
-                        composer.push_layer(Layer::Offset(OffsetLayer::new(origin + child_offset)));
-                        self.paint_subtree(composer, child_id, Offset::ZERO, dirty_set)?;
-                        composer.pop_layer();
-                    } else {
-                        // Inline children bake into the shared picture
-                        // space — runs merge, no extra layer.
-                        self.paint_subtree(composer, child_id, origin + child_offset, dirty_set)?;
-                    }
-                }
-            }
-        }
-
-        for _ in 0..effect_layers {
-            composer.pop_layer();
-        }
-
-        Ok(())
-    }
-}
-
-// ============================================================================
-// Fragment composition (paint phase plumbing)
-// ============================================================================
-
-/// Builds the frame's [`LayerTree`] from replayed paint fragments,
-/// merging adjacent inline draw runs into shared `PictureLayer`s.
-///
-/// Sealing discipline mirrors the recorder's: the open run is flushed
-/// into a `PictureLayer` whenever a layer boundary needs ordering
-/// (push/pop) and at [`Self::finish`]. The stack always holds at least
-/// the root `OffsetLayer`.
-#[derive(Debug)]
-struct FragmentComposer {
-    tree: LayerTree,
-    stack: Vec<LayerId>,
-    open: DisplayList,
-}
-
-impl FragmentComposer {
-    /// `device_pixel_ratio` becomes the root layer's scale: the
-    /// framework paints in LOGICAL pixels, the engine rasterizes in
-    /// physical surface pixels — the root transform is the single
-    /// place the two meet (Flutter's RenderView root transform).
-    fn new(device_pixel_ratio: f32) -> Self {
-        let mut tree = LayerTree::new();
-        let root_layer = if (device_pixel_ratio - 1.0).abs() < f32::EPSILON {
-            Layer::Offset(OffsetLayer::zero())
-        } else {
-            Layer::Transform(TransformLayer::new(flui_types::Matrix4::scaling(
-                device_pixel_ratio,
-                device_pixel_ratio,
-                1.0,
-            )))
-        };
-        let root = tree.insert(root_layer);
-        tree.set_root(Some(root));
-        Self {
-            tree,
-            stack: vec![root],
-            open: DisplayList::new(),
-        }
-    }
-
-    /// Merges a sealed fragment run into the open picture.
-    fn append_run(&mut self, run: DisplayList) {
-        self.open.append(run);
-    }
-
-    /// Flushes the open picture into a `PictureLayer` under the
-    /// current stack top (no-op when empty).
-    fn seal_picture(&mut self) {
-        if flui_painting::DisplayListCore::is_empty(&self.open) {
-            return;
-        }
-        let list = std::mem::take(&mut self.open);
-        let layer_id = self.tree.insert(Layer::from(PictureLayer::new(list)));
-        let parent = *self
-            .stack
-            .last()
-            .expect("composer stack always holds the root layer (popping it is rejected)");
-        self.tree.add_child(parent, layer_id);
-    }
-
-    fn push_layer(&mut self, layer: Layer) {
-        self.seal_picture();
-        let id = self.tree.insert(layer);
-        let parent = *self
-            .stack
-            .last()
-            .expect("composer stack always holds the root layer (popping it is rejected)");
-        self.tree.add_child(parent, id);
-        self.stack.push(id);
-    }
-
-    fn pop_layer(&mut self) {
-        self.seal_picture();
-        debug_assert!(
-            self.stack.len() > 1,
-            "composer pop without matching push — fragment scope ops are \
-             balanced by the recorder, so an underflow means the replay \
-             loop pushed/popped asymmetrically",
-        );
-        if self.stack.len() > 1 {
-            self.stack.pop();
-        }
-    }
-
-    fn finish(mut self) -> LayerTree {
-        self.seal_picture();
-        debug_assert_eq!(
-            self.stack.len(),
-            1,
-            "composer finished with unbalanced layer stack — every \
-             push_layer in the replay loop must have a matching pop_layer",
-        );
-        self.tree
-    }
-}
-
-/// Maps a recorded clip scope onto its `flui-layer` clip layer.
-///
-/// Clip shapes are recorded in the node's LOCAL coordinates, while the
-/// runs they bracket carry the accumulated `origin` baked into their
-/// canvas transforms — so the shape is shifted by `origin` here
-/// (Flutter `pushClipRect`: `clipRect.shift(offset)`), or a clip away
-/// from the parent origin would cut at the layer's (0,0) instead of
-/// the node's position.
-///
-/// Always a real clip layer today; lowering non-composited clips back
-/// into canvas clips inside the merged picture is a composer-side
-/// optimization gated on the `needs_compositing` bits — correctness is
-/// identical either way, so the recording API does not expose the
-/// choice.
-fn clip_layer(clip: FragmentClip, origin: Offset) -> Layer {
-    match clip {
-        FragmentClip::Rect { rect, behavior } => {
-            Layer::ClipRect(ClipRectLayer::new(rect.translate_offset(origin), behavior))
-        }
-        FragmentClip::RRect { rrect, behavior } => Layer::ClipRRect(ClipRRectLayer::new(
-            rrect.translate_offset(origin),
-            behavior,
-        )),
-        FragmentClip::Path { path, behavior } => {
-            let path = if origin == Offset::ZERO {
-                *path
-            } else {
-                path.translate(origin)
-            };
-            Layer::ClipPath(Box::new(ClipPathLayer::new(path, behavior)))
-        }
-    }
-}
-
-// ============================================================================
-// Semantics phase: run_semantics
-// ============================================================================
-
-impl PipelineOwner<Semantics> {
-    /// Completes the frame and returns to [`Idle`].
-    #[must_use]
-    pub fn finish(self) -> PipelineOwner<Idle> {
-        rebind_phase(self)
-    }
-
-    /// Updates semantics for all dirty render objects.
-    ///
-    /// This is phase 4 of the rendering pipeline. During semantics:
-    /// - Accessibility information is gathered
-    /// - Semantics tree is updated
-    ///
-    /// Nodes are sorted by depth (shallow first) for top-down traversal.
-    /// The geometries of children depend on ancestors' transforms and clips,
-    /// so parents must be processed first. This matches Flutter's
-    /// `flushSemantics`.
-    pub fn run_semantics(&mut self) -> crate::error::RenderResult<()> {
-        if !self.semantics_enabled() {
-            return Ok(());
-        }
-
-        tracing::debug!(
-            "run_semantics: {} nodes",
-            self.scheduler.semantics_queue_len()
-        );
-
-        self.scheduler.enter_phase(PhaseKind::Semantics);
-
-        // PR #109 review feedback: pre-fix this path used
-        // `std::mem::take(&mut self.dirty.needs_semantics)` to drain in
-        // one step. Take leaves an empty `Vec::new()` (capacity 0)
-        // behind, so every subsequent semantics-enabled frame's first
-        // push re-allocates. Switch to an in-place sort + iterate +
-        // clear pattern that preserves the Vec's backing capacity
-        // across frames (idiom: *Programming Rust* 2nd ed §11 "Owned
-        // vs Borrowed", retain the allocation by retaining ownership).
-        // The Flutter-parity `where !object._needsLayout` filter the
-        // pre-cycle comment promised was never implemented; that gap
-        // lands when the real semantics-config build is wired (R-1
-        // follow-up).
-
-        // Sort shallow-first matching Flutter's flushSemantics. Roots
-        // dispatch before their descendants so a parent's config is
-        // assembled before children fold into it.
-        self.scheduler.sort_semantics_shallow_first();
-
-        // Cycle 4 R-1: pre-cycle the path panicked with
-        // `unimplemented!()` once any node was queued — a Constitution
-        // Principle 6 violation in a hot-path callable from
-        // `RendererBinding::draw_frame` on every frame as soon as
-        // semantics_enabled() flipped true.
-        //
-        // Post-cycle: walk the dirty list, emit a `tracing::warn!`
-        // per node carrying the missing-integration hint, and return
-        // `Ok(())`. The framework no longer aborts on semantics flips;
-        // when the full `SemanticsOwner` integration lands, swap the
-        // warn for the real config-build + owner-register call.
-        //
-        // Aggregate into a count rather than emitting one warn per node
-        // (avoids O(n) log spam when semantics is enabled on a large tree).
-        // `nodes_needing_semantics()` is a shared slice accessor — disjoint
-        // from the mutable `self.render_tree` field under Rust 2024 capture.
-        let pending_count = self
-            .scheduler
-            .nodes_needing_semantics()
-            .iter()
-            .filter(|d| self.render_tree.contains(d.id))
-            .count();
-        if pending_count > 0 {
-            tracing::warn!(
-                count = pending_count,
-                "run_semantics: full SemanticsOwner integration pending; \
-                 semantics config build for {pending_count} node(s) is a no-op until \
-                 RenderObject → SemanticsConfiguration plumbing lands"
-            );
-        }
-        // `clear()` retains the Vec's allocated capacity; next frame's
-        // pushes amortise into the existing buffer.
-        self.scheduler.clear_semantics_queue();
-
-        // exit_phase clears debug_doing_semantics AND drains mid-semantics
-        // marks so semantics marks made during this iteration's
-        // `debug_doing_semantics = true` window aren't stranded. Drained
-        // entries land on dirty.needs_semantics for the NEXT run_semantics.
-        let _ = self.scheduler.exit_phase(PhaseKind::Semantics);
-
-        Ok(())
-    }
-}
-
 /// Internal helper: shifts the `Phase` phantom parameter without touching any
 /// runtime field. Behaviour-preserving by construction.
 #[inline]
@@ -2981,376 +234,6 @@ where
         #[cfg(any(test, feature = "testing"))]
         parent_data_seeds: from.parent_data_seeds,
         _phase: PhantomData,
-    }
-}
-
-// ============================================================================
-// Memoizing query walks (intrinsics / dry layout)
-// ============================================================================
-
-/// One node's slot in a memoizing query walk: the disjoint `&mut`
-/// borrow plus a snapshot of the node's child ids. The node is moved
-/// OUT (`node.take()`) while its own computation runs, so re-entry —
-/// which only a cyclic child link can produce — is detected instead of
-/// aliasing the borrow.
-struct QuerySlot<'a> {
-    node: Option<&'a mut crate::storage::RenderNode>,
-    children: Vec<RenderId>,
-}
-
-/// Recursive memoized intrinsic query over the take-out slot map.
-///
-/// Per node: cache peek → on miss, run the object's `intrinsic_raw`
-/// with a child callback that recurses through this same function →
-/// store the result. Errors inside the child callback are stashed and
-/// re-raised after the object call returns (the raw callback channel
-/// is infallible by design — same convention as the hit-test walk).
-fn intrinsic_query(
-    slots: &mut rustc_hash::FxHashMap<RenderId, QuerySlot<'_>>,
-    id: RenderId,
-    dimension: crate::storage::IntrinsicDimension,
-    extent: f32,
-    #[cfg(any(test, feature = "testing"))] parent_data_seeds: &FxHashMap<RenderId, ParentDataSeed>,
-    #[cfg(not(any(test, feature = "testing")))] parent_data_seeds: &(),
-) -> crate::error::RenderResult<f32> {
-    ensure_stack(|| intrinsic_query_impl(slots, id, dimension, extent, parent_data_seeds))
-}
-
-/// Body of [`intrinsic_query`]; split out so every recursion level
-/// enters through the [`ensure_stack`] probe.
-fn intrinsic_query_impl(
-    slots: &mut rustc_hash::FxHashMap<RenderId, QuerySlot<'_>>,
-    id: RenderId,
-    dimension: crate::storage::IntrinsicDimension,
-    extent: f32,
-    #[cfg(any(test, feature = "testing"))] parent_data_seeds: &FxHashMap<RenderId, ParentDataSeed>,
-    #[cfg(not(any(test, feature = "testing")))] parent_data_seeds: &(),
-) -> crate::error::RenderResult<f32> {
-    let Some(slot) = slots.get_mut(&id) else {
-        return Err(crate::error::RenderError::NodeNotFound(id));
-    };
-    let Some(node) = slot.node.take() else {
-        // Only reachable through a cyclic child link: the node's own
-        // computation is still on the stack. Degenerate-but-defined in
-        // release; loud in debug (collect_subtree_ids already refuses
-        // to loop, so the cycle must close through duplicate child
-        // indices).
-        debug_assert!(
-            false,
-            "intrinsic query re-entered node {id:?} mid-computation — cyclic child links"
-        );
-        return Ok(0.0);
-    };
-    let children = slot.children.clone();
-
-    let result = (|| {
-        let Some(entry) = node.as_box_mut() else {
-            return Err(crate::error::RenderError::ProtocolMismatch {
-                node_protocol: "sliver",
-                constraints_protocol: "box",
-            });
-        };
-        if let Some(hit) = entry
-            .state()
-            .layout_cache()
-            .peek_intrinsic(dimension, extent)
-        {
-            return Ok(hit);
-        }
-        let mut child_err: Option<crate::error::RenderError> = None;
-        let value = {
-            let child_err = &mut child_err;
-            let mut child_query =
-                |index: usize, dim: crate::storage::IntrinsicDimension, ext: f32| -> f32 {
-                    let Some(&child_id) = children.get(index) else {
-                        child_err.get_or_insert(crate::error::RenderError::contract_violation(
-                            "intrinsic child query",
-                            "child index out of range for this node's children",
-                        ));
-                        return 0.0;
-                    };
-                    match intrinsic_query(slots, child_id, dim, ext, parent_data_seeds) {
-                        Ok(v) => v,
-                        Err(err) => {
-                            child_err.get_or_insert(err);
-                            0.0
-                        }
-                    }
-                };
-            let mut child_flex = |index: usize| -> i32 {
-                child_flex_from_seeds(parent_data_seeds, &children, index)
-            };
-            entry.render_object().intrinsic_raw(
-                dimension,
-                extent,
-                children.len(),
-                &mut child_query,
-                &mut child_flex,
-            )
-        };
-        if let Some(err) = child_err {
-            return Err(err);
-        }
-        entry
-            .state_mut()
-            .layout_cache_mut()
-            .insert_intrinsic(dimension, extent, value);
-        Ok(value)
-    })();
-
-    // Restore the slot even on the error path — sibling queries in the
-    // same walk must still find the node.
-    if let Some(slot) = slots.get_mut(&id) {
-        slot.node = Some(node);
-    }
-    result
-}
-
-/// Recursive memoized dry-layout query; same skeleton as
-/// [`intrinsic_query`] with `(constraints → Size)` payloads.
-fn dry_layout_query(
-    slots: &mut rustc_hash::FxHashMap<RenderId, QuerySlot<'_>>,
-    id: RenderId,
-    constraints: crate::constraints::BoxConstraints,
-) -> crate::error::RenderResult<flui_types::Size> {
-    ensure_stack(|| dry_layout_query_impl(slots, id, constraints))
-}
-
-/// Body of [`dry_layout_query`]; split out so every recursion level
-/// enters through the [`ensure_stack`] probe.
-fn dry_layout_query_impl(
-    slots: &mut rustc_hash::FxHashMap<RenderId, QuerySlot<'_>>,
-    id: RenderId,
-    constraints: crate::constraints::BoxConstraints,
-) -> crate::error::RenderResult<flui_types::Size> {
-    let Some(slot) = slots.get_mut(&id) else {
-        return Err(crate::error::RenderError::NodeNotFound(id));
-    };
-    let Some(node) = slot.node.take() else {
-        debug_assert!(
-            false,
-            "dry-layout query re-entered node {id:?} mid-computation — cyclic child links"
-        );
-        return Ok(flui_types::Size::ZERO);
-    };
-    let children = slot.children.clone();
-
-    let result = (|| {
-        let Some(entry) = node.as_box_mut() else {
-            return Err(crate::error::RenderError::ProtocolMismatch {
-                node_protocol: "sliver",
-                constraints_protocol: "box",
-            });
-        };
-        if let Some(hit) = entry.state().layout_cache().peek_dry_layout(constraints) {
-            return Ok(hit);
-        }
-        let mut child_err: Option<crate::error::RenderError> = None;
-        let value = {
-            let child_err = &mut child_err;
-            let mut child_dry =
-                |index: usize, c: crate::constraints::BoxConstraints| -> flui_types::Size {
-                    let Some(&child_id) = children.get(index) else {
-                        child_err.get_or_insert(crate::error::RenderError::contract_violation(
-                            "dry-layout child query",
-                            "child index out of range for this node's children",
-                        ));
-                        return flui_types::Size::ZERO;
-                    };
-                    match dry_layout_query(slots, child_id, c) {
-                        Ok(v) => v,
-                        Err(err) => {
-                            child_err.get_or_insert(err);
-                            flui_types::Size::ZERO
-                        }
-                    }
-                };
-            entry
-                .render_object()
-                .dry_layout_raw(constraints, children.len(), &mut child_dry)
-        };
-        if let Some(err) = child_err {
-            return Err(err);
-        }
-        entry
-            .state_mut()
-            .layout_cache_mut()
-            .insert_dry_layout(constraints, value);
-        Ok(value)
-    })();
-
-    if let Some(slot) = slots.get_mut(&id) {
-        slot.node = Some(node);
-    }
-    result
-}
-
-/// Recursive memoized dry-baseline query; same skeleton as
-/// [`dry_layout_query`] with `(constraints, baseline → Option<f32>)`
-/// payloads.
-fn dry_baseline_query(
-    slots: &mut rustc_hash::FxHashMap<RenderId, QuerySlot<'_>>,
-    id: RenderId,
-    constraints: crate::constraints::BoxConstraints,
-    baseline: crate::traits::TextBaseline,
-) -> crate::error::RenderResult<Option<f32>> {
-    ensure_stack(|| dry_baseline_query_impl(slots, id, constraints, baseline))
-}
-
-/// Body of [`dry_baseline_query`]; split out so every recursion level
-/// enters through the [`ensure_stack`] probe.
-fn dry_baseline_query_impl(
-    slots: &mut rustc_hash::FxHashMap<RenderId, QuerySlot<'_>>,
-    id: RenderId,
-    constraints: crate::constraints::BoxConstraints,
-    baseline: crate::traits::TextBaseline,
-) -> crate::error::RenderResult<Option<f32>> {
-    let Some(slot) = slots.get_mut(&id) else {
-        return Err(crate::error::RenderError::NodeNotFound(id));
-    };
-    let Some(node) = slot.node.take() else {
-        debug_assert!(
-            false,
-            "dry-baseline query re-entered node {id:?} mid-computation — cyclic child links"
-        );
-        return Ok(None);
-    };
-    let children = slot.children.clone();
-
-    let result = (|| {
-        let Some(entry) = node.as_box_mut() else {
-            return Err(crate::error::RenderError::ProtocolMismatch {
-                node_protocol: "sliver",
-                constraints_protocol: "box",
-            });
-        };
-        if let Some(hit) = entry
-            .state()
-            .layout_cache()
-            .peek_dry_baseline(constraints, baseline)
-        {
-            return Ok(hit);
-        }
-        let mut child_err: Option<crate::error::RenderError> = None;
-        let value = {
-            let child_err = &mut child_err;
-            let mut child_query = |index: usize,
-                                   request: crate::context::DryBaselineChildRequest|
-             -> crate::context::DryBaselineChildResponse {
-                use crate::context::{DryBaselineChildRequest, DryBaselineChildResponse};
-                let Some(&child_id) = children.get(index) else {
-                    child_err.get_or_insert(crate::error::RenderError::contract_violation(
-                        "dry-baseline child query",
-                        "child index out of range for this node's children",
-                    ));
-                    return match request {
-                        DryBaselineChildRequest::Baseline(_, _) => {
-                            DryBaselineChildResponse::Baseline(None)
-                        }
-                        DryBaselineChildRequest::DryLayout(_) => {
-                            DryBaselineChildResponse::DryLayout(flui_types::Size::ZERO)
-                        }
-                    };
-                };
-                match request {
-                    DryBaselineChildRequest::Baseline(c, b) => {
-                        match dry_baseline_query(slots, child_id, c, b) {
-                            Ok(v) => DryBaselineChildResponse::Baseline(v),
-                            Err(err) => {
-                                child_err.get_or_insert(err);
-                                DryBaselineChildResponse::Baseline(None)
-                            }
-                        }
-                    }
-                    DryBaselineChildRequest::DryLayout(c) => {
-                        match dry_layout_query(slots, child_id, c) {
-                            Ok(v) => DryBaselineChildResponse::DryLayout(v),
-                            Err(err) => {
-                                child_err.get_or_insert(err);
-                                DryBaselineChildResponse::DryLayout(flui_types::Size::ZERO)
-                            }
-                        }
-                    }
-                }
-            };
-            entry.render_object().dry_baseline_raw(
-                constraints,
-                baseline,
-                children.len(),
-                &mut child_query,
-            )
-        };
-        if let Some(err) = child_err {
-            return Err(err);
-        }
-        entry
-            .state_mut()
-            .layout_cache_mut()
-            .insert_dry_baseline(constraints, baseline, value);
-        Ok(value)
-    })();
-
-    if let Some(slot) = slots.get_mut(&id) {
-        slot.node = Some(node);
-    }
-    result
-}
-
-// ============================================================================
-// Diagnostics dump (Diagnosticable-backed)
-// ============================================================================
-
-/// Builds a [`DiagnosticsNode`] for a single render node: the render object
-/// self-describes its own properties, and the committed geometry/offset is
-/// layered on from `RenderState`.
-fn node_diagnostics(node: &RenderNode) -> DiagnosticsNode {
-    if let Some(entry) = node.as_box() {
-        let mut diagnostics = entry.render_object().to_diagnostics_node();
-        diagnostics = diagnostics.property("paint_offset", format!("{:?}", node.offset()));
-        if let Some(size) = entry.state().geometry() {
-            diagnostics = diagnostics.property("size", format!("{size:?}"));
-        }
-        diagnostics
-    } else if let Some(entry) = node.as_sliver() {
-        let mut diagnostics = entry.render_object().to_diagnostics_node();
-        diagnostics = diagnostics.property("paint_offset", format!("{:?}", node.offset()));
-        if let Some(geometry) = entry.state().geometry() {
-            diagnostics = diagnostics.property("geometry", format!("{geometry:?}"));
-        }
-        diagnostics
-    } else {
-        DiagnosticsNode::new("<unknown>")
-    }
-}
-
-impl<Phase: PipelinePhase> PipelineOwner<Phase> {
-    /// Returns a [`DiagnosticsNode`] tree mirroring the render hierarchy
-    /// rooted at [`root_id`](Self::root_id), or `None` if no root is set.
-    ///
-    /// Each node self-describes via its render object's
-    /// [`Diagnosticable`](flui_foundation::Diagnosticable) impl, with the
-    /// committed geometry/offset layered on. This is the
-    /// `Diagnosticable`-backed counterpart to a plain `{:?}` dump and the
-    /// basis for [`debug_dump_pipeline_owner_tree`](crate::binding::debug_dump_pipeline_owner_tree).
-    pub fn debug_diagnostics_tree(&self) -> Option<DiagnosticsNode> {
-        self.debug_diagnostics_subtree(self.root_id?)
-    }
-
-    /// Returns the [`DiagnosticsNode`] for a single node `id` (no children),
-    /// or `None` if the id is not live.
-    pub fn debug_node_diagnostics(&self, id: RenderId) -> Option<DiagnosticsNode> {
-        Some(node_diagnostics(self.render_tree.get(id)?))
-    }
-
-    fn debug_diagnostics_subtree(&self, id: RenderId) -> Option<DiagnosticsNode> {
-        let node = self.render_tree.get(id)?;
-        let mut diagnostics = node_diagnostics(node);
-        for &child in node.children() {
-            if let Some(child_diagnostics) = self.debug_diagnostics_subtree(child) {
-                diagnostics.add_child(child_diagnostics);
-            }
-        }
-        Some(diagnostics)
     }
 }
